refactor(parser): split expr/precedence.rs into cohesive submodules (§7-8)
Split the 2961-line parser/expr/precedence.rs (over the 500-line limit)
into a facade plus twelve themed submodules under
src/parser/expr/precedence/. All but one are under 500 lines;
list_infix_loop.rs (506) holds the single indivisible parse_list_infix_loop
function, which cannot be split at a helper boundary. precedence.rs is free
parser functions addressed as precedence::<name> (11 call sites); the
facade re-exports each submodule preserving its original visibility, so no
external caller changes. expr/mod.rs is untouched.

- precedence.rs (88)        - facade: use lines, CHAIN_CMP_TMP_COUNTER, re-exports
- list_infix_loop.rs (506)  - parse_list_infix_loop (single indivisible fn)
- comparison.rs (419)       - comparison_expr_mode + regex_rhs_needs_more_parsing
- logic2.rs (312)           - not/flipflop/or-or/and-and/junctive modes
- assign.rs (302)           - assignment-target lvalue builders
- chain_cmp.rs (281)        - chain comparison + range_expr + smartmatch rhs
- logic.rs (269)            - or/and expr modes
- ternary.rs (259)          - ternary + call_arg_expr + structural comparison
- feed.rs (247)             - feed-operator (==> / <==) lowering
- custom_infix.rs (184)     - custom/flipflop infix + comma-list-of-range
- errors.rs (75)            - precedence/associativity error builders
- list_infix.rs (66)        - list_infix_expr / sequence_expr / maybe_wrap_lhs
- operands.rs (21)          - loose-prefix/multiplicative/power operands

Pure relocation. Moved fns are pub(crate) in submodules (four taking
pub(super) operator-enum types are pub(in crate::parser::expr) to avoid
private-interface leaks) and re-exported at their original visibility.
Body refs to expr-level siblings rewritten to absolute crate::parser paths.
The fn multiset is byte-identical. No behavior change.
