+++ src/parser/expr/precedence.rs
@@ -22,2929 +22,56 @@ use super::{contains_whatever, wrap_whatevercode};
 
 static CHAIN_CMP_TMP_COUNTER: AtomicUsize = AtomicUsize::new(0);
 
-fn syntax_exception(class_name: &str, message: impl Into<String>) -> PError {
-    let message = message.into();
-    let mut attrs = HashMap::new();
-    attrs.insert("message".to_string(), Value::str(message.clone()));
-    let exception = Value::make_instance(Symbol::intern(class_name), attrs);
-    PError::fatal_with_exception(message, Box::new(exception))
-}
-
-fn worry_precedence_range(action: &str) -> PError {
-    let message = format!(
-        "To {} a range, parenthesize the whole range.\n(Or parenthesize the whole endpoint expression, if you meant that.)",
-        action
-    );
-    let mut attrs = HashMap::new();
-    attrs.insert("message".to_string(), Value::str(message.clone()));
-    attrs.insert("action".to_string(), Value::str(action.to_string()));
-    let exception = Value::make_instance(Symbol::intern("X::Worry::Precedence::Range"), attrs);
-    PError::fatal_with_exception(message, Box::new(exception))
-}
-
-/// Check if the left expression of a range operator is a bare prefix `|` or `~`
-/// (not parenthesized), which creates a precedence ambiguity warning.
-/// We check the original input text rather than the AST because parentheses
-/// are transparent in the AST (`(|4)` and `|4` produce the same tree).
-fn check_range_precedence_worry(input: &str) -> Result<(), PError> {
-    let trimmed = input.trim_start();
-    if trimmed.starts_with('|') && !trimmed.starts_with("|(") {
-        return Err(worry_precedence_range("apply a Slip flattener to"));
-    }
-    if trimmed.starts_with('~') && !trimmed.starts_with("~(") {
-        return Err(worry_precedence_range("stringify"));
-    }
-    Ok(())
-}
-
-fn non_list_associative_error(lhs: &str, rhs: &str) -> PError {
-    syntax_exception(
-        "X::Syntax::NonListAssociative",
-        format!(
-            "Only identical operators may be list associative; since '{}' and '{}' differ, they are non-associative and you need to clarify with parentheses",
-            lhs, rhs
-        ),
-    )
-}
-
-fn non_associative_error(op_name: &str) -> PError {
-    syntax_exception(
-        "X::Syntax::NonAssociative",
-        format!("Non-associative operator '{}' cannot be chained", op_name),
-    )
-}
-
-/// Non-associative chain of two named operators (e.g. `1 <=> 2 <=> 3`).
-/// Carries `.left` / `.right` like rakudo's `X::Syntax::NonAssociative`.
-fn non_associative_pair_error(left: &str, right: &str) -> PError {
-    let message = format!(
-        "Operators '{}' and '{}' are non-associative and require parentheses",
-        left, right
-    );
-    let mut attrs = HashMap::new();
-    attrs.insert("message".to_string(), Value::str(message.clone()));
-    attrs.insert("left".to_string(), Value::str(left.to_string()));
-    attrs.insert("right".to_string(), Value::str(right.to_string()));
-    let exception = Value::make_instance(Symbol::intern("X::Syntax::NonAssociative"), attrs);
-    PError::fatal_with_exception(message, Box::new(exception))
-}
-
-fn conditional_precedence_too_loose_error() -> PError {
-    syntax_exception(
-        "X::Syntax::ConditionalOperator::PrecedenceTooLoose",
-        "Assignment operators inside ?? !! are too loose; parenthesize them",
-    )
-}
-
-fn is_assignment_expr(expr: &Expr) -> bool {
-    matches!(
-        expr,
-        Expr::AssignExpr { .. } | Expr::IndexAssign { .. } | Expr::MultiDimIndexAssign { .. }
-    )
-}
-
-fn comparison_nonassoc_key(op: &TokenKind) -> Option<&'static str> {
-    match op {
-        TokenKind::LtEqGt => Some("<=>"),
-        TokenKind::Ident(name) if name == "leg" => Some("leg"),
-        TokenKind::Ident(name) if name == "cmp" => Some("cmp"),
-        TokenKind::Ident(name) if name == "coll" => Some("coll"),
-        TokenKind::Ident(name) if name == "unicmp" => Some("unicmp"),
-        // eqv, before, after are chaining operators (not non-associative)
-        _ => None,
-    }
-}
-
-fn is_structural_comparison_op(op: ComparisonOp) -> bool {
-    // Only truly non-associative operators that return Order (not Bool).
-    // eqv, before, after are chaining operators and handled by the
-    // regular comparison chaining path.
-    matches!(
-        op,
-        ComparisonOp::Spaceship
-            | ComparisonOp::Leg
-            | ComparisonOp::Cmp
-            | ComparisonOp::Coll
-            | ComparisonOp::Unicmp
-    )
-}
-
-fn structural_comparison_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (rest, left) = junctive_expr_mode(input, mode)?;
-    let (r, _) = ws(rest)?;
-    let Some((op, len)) = parse_comparison_op(r) else {
-        return Ok((rest, left));
-    };
-    if !is_structural_comparison_op(op) {
-        return Ok((rest, left));
-    }
-    let r = &r[len..];
-    let (r, _) = ws(r)?;
-    let (r, right) = if mode == ExprMode::Full {
-        junctive_expr_mode(r, mode).map_err(|err| PError {
-            messages: merge_expected_messages(
-                "expected expression after comparison operator",
-                &err.messages,
-            ),
-            remaining_len: err.remaining_len.or(Some(r.len())),
-            exception: None,
-        })?
-    } else {
-        junctive_expr_mode(r, mode)?
-    };
-    let expr = Expr::Binary {
-        left: Box::new(left),
-        op: op.token_kind(),
-        right: Box::new(right),
-    };
-    let (r2, _) = ws(r)?;
-    if let Some((next_op, _)) = parse_comparison_op(r2)
-        && is_structural_comparison_op(next_op)
-    {
-        return Err(non_associative_pair_error(
-            comparison_nonassoc_key(&op.token_kind()).unwrap_or("<=>"),
-            comparison_nonassoc_key(&next_op.token_kind()).unwrap_or("<=>"),
-        ));
-    }
-    Ok((r, expr))
-}
-
-/// Ternary: expr ?? expr !! expr
-pub(super) fn ternary(input: &str) -> PResult<'_, Expr> {
-    ternary_mode(input, ExprMode::Full)
-}
-
-/// Parse a no-paren callable argument expression.
-///
-/// This uses the precedence tier just tighter than loose word-logicals
-/// (`and`, `or`, `xor`, `orelse`, `andthen`, `notandthen`) so that:
-/// - `foo 3 != 3` parses as `foo(3 != 3)`
-/// - `isfive 5 and isfive 5` parses as `isfive(5) and isfive(5)`
-///
-/// Also handles item assignment (`=`, `~=`, `+=`, etc.) within a single
-/// argument so that `f $a ~= $b, $c` parses as `f(($a ~= $b), $c)`.
-pub(in crate::parser) fn call_arg_expr(input: &str) -> PResult<'_, Expr> {
-    let (rest, expr) = call_arg_ternary_expr(input)?;
-    let (r, _) = ws(rest)?;
-
-    // Handle compound assignment operators (+=, ~=, //=, etc.) in call-arg context.
-    // The RHS is a single expression (no comma collection) since commas separate
-    // function arguments at this level. Item assignment is looser than the
-    // conditional `?? !!`, so the RHS parses a ternary too.
-    if let Some((after_op, op)) = parse_compound_assign_op(r) {
-        let (after_ws, _) = ws(after_op)?;
-        if let Ok((r2, rhs)) = call_arg_ternary_expr(after_ws)
-            && let Ok(result) = build_compound_assign_expr(expr.clone(), op, rhs)
-        {
-            return Ok((r2, result));
-        }
-    }
-
-    // Handle simple assignment (=) in call-arg context.
-    if r.starts_with('=') && !r.starts_with("==") && !r.starts_with("=>") {
-        let r_eq = &r[1..];
-        let (r_eq, _) = ws(r_eq)?;
-        if let Ok((r2, rhs)) = call_arg_ternary_expr(r_eq) {
-            return Ok((r2, assign_to_target_expr(expr, rhs)));
-        }
-    }
-
-    Ok((rest, expr))
-}
-
-/// Parse a single call argument at *conditional* (`?? !!`) precedence: an
-/// `or_or` expression optionally followed by a ternary. The comma operator
-/// separates arguments at a looser level, so `f $cond ?? $a !! $b` binds the
-/// ternary inside the argument (matching Raku) rather than wrapping the whole
-/// call as `(f $cond) ?? $a !! $b`.
-fn call_arg_ternary_expr(input: &str) -> PResult<'_, Expr> {
-    let (rest, cond) = or_or_expr_mode(input, ExprMode::NoSequenceNoFeed)?;
-    let (r, _) = ws(rest)?;
-    let Ok((r, _)) = parse_tag(r, "??") else {
-        return Ok((rest, cond));
-    };
-    let (r, _) = ws(r)?;
-    let (r, then_expr) = call_arg_ternary_expr(r).map_err(|err| {
-        enrich_expected_error(err, "expected then-expression after '??'", r.len())
-    })?;
-    let (r, _) = ws(r)?;
-    let (r, _) = parse_tag(r, "!!").map_err(|err| {
-        enrich_expected_error(err, "expected '!!' in ternary expression", r.len())
-    })?;
-    let (r, _) = ws(r)?;
-    let (r, else_expr) = call_arg_ternary_expr(r).map_err(|err| {
-        enrich_expected_error(err, "expected else-expression after '!!'", r.len())
-    })?;
-    Ok((
-        r,
-        Expr::Ternary {
-            cond: Box::new(cond),
-            then_expr: Box::new(then_expr),
-            else_expr: Box::new(else_expr),
-        },
-    ))
-}
-
-/// Parse a ternary-level expression that does NOT consume a trailing simple or
-/// compound assignment (`=`, `+=`, ...). Used for signature `where` constraints,
-/// where a following `=` introduces the parameter default rather than an
-/// assignment to the constraint expression. `ternary_mode` cannot be reused here
-/// because its no-`??` fallback (`or_expr_mode`) includes assignment, which would
-/// swallow the default value (e.g. `$x where Int = 9`).
-pub(super) fn ternary_no_assign(input: &str) -> PResult<'_, Expr> {
-    let (rest, cond) = or_expr_no_assign_mode(input, ExprMode::Full)?;
-    let (rest_ws, _) = ws(rest)?;
-    if rest_ws.starts_with("??") {
-        // A conditional constraint: the full ternary parser is safe because a
-        // top-level `=` only appears after the complete `?? !!` expression.
-        return ternary_mode(input, ExprMode::Full);
-    }
-    Ok((rest, cond))
-}
-
-pub(super) fn ternary_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (rest, cond) = or_expr_no_assign_mode(input, mode)?;
-    let (rest_ws, _) = ws(rest)?;
-    if let Ok((input, _)) = parse_tag(rest_ws, "??") {
-        let (input, _) = ws(input)?;
-        // Parse then-expr at ternary precedence (stops before `!!`)
-        let (input, then_expr) = if mode == ExprMode::Full {
-            ternary_mode(input, mode).map_err(|err| {
-                enrich_expected_error(err, "expected then-expression after '??'", input.len())
-            })?
-        } else {
-            ternary_mode(input, mode)?
-        };
-        if mode == ExprMode::Full
-            && let Expr::BareWord(name) = &then_expr
-            && !name.contains("::")
-        {
-            return Err(PError::expected("expected '!!' in ternary expression"));
-        }
-        let (input, _) = ws(input)?;
-        let (input, _) = if mode == ExprMode::Full {
-            parse_tag(input, "!!").map_err(|err| {
-                enrich_expected_error(err, "expected '!!' in ternary expression", input.len())
-            })?
-        } else {
-            parse_tag(input, "!!")?
-        };
-        let (input, _) = ws(input)?;
-        // Parse else-expr at ternary precedence (allows nested ternary)
-        let (input, else_expr) = if mode == ExprMode::Full {
-            ternary_mode(input, mode).map_err(|err| {
-                enrich_expected_error(err, "expected else-expression after '!!'", input.len())
-            })?
-        } else {
-            ternary_mode(input, mode)?
-        };
-        if is_assignment_expr(&then_expr) || is_assignment_expr(&else_expr) {
-            return Err(conditional_precedence_too_loose_error());
-        }
-        if let Expr::Binary { left, op, right } = cond {
-            // The loose word-logicals (`and`/`andthen`/`notandthen`/`or`/`orelse`)
-            // are LOOSER than the conditional `?? !!`, so when one was greedily
-            // folded into the condition the ternary must re-associate to bind
-            // tighter (its right operand). The tight `&&`/`||` are TIGHTER than
-            // `?? !!`, so they must NOT re-associate: `$a && $b ?? $c !! $d` is
-            // `($a && $b) ?? $c !! $d`, handled by the fall-through below. The
-            // loose `and` carries its own `AndWord` token precisely so it can be
-            // distinguished here from the tight `&&` (`AndAnd`).
-            if matches!(
-                op,
-                TokenKind::AndWord
-                    | TokenKind::AndThen
-                    | TokenKind::NotAndThen
-                    | TokenKind::OrWord
-                    | TokenKind::OrElse
-            ) {
-                return Ok((
-                    input,
-                    Expr::Binary {
-                        left,
-                        op,
-                        right: Box::new(Expr::Ternary {
-                            cond: right,
-                            then_expr: Box::new(then_expr),
-                            else_expr: Box::new(else_expr),
-                        }),
-                    },
-                ));
-            }
-            return Ok((
-                input,
-                Expr::Ternary {
-                    cond: Box::new(Expr::Binary { left, op, right }),
-                    then_expr: Box::new(then_expr),
-                    else_expr: Box::new(else_expr),
-                },
-            ));
-        }
-        return Ok((
-            input,
-            Expr::Ternary {
-                cond: Box::new(cond),
-                then_expr: Box::new(then_expr),
-                else_expr: Box::new(else_expr),
-            },
-        ));
-    }
-    // No ternary operator: fall back to regular OR-level parsing, which includes
-    // assignment expressions.
-    or_expr_mode(input, mode)
-}
-
-/// Parse an operand at additive level or tighter (for loose prefix operators).
-pub(super) fn loose_prefix_operand(input: &str, level: i32) -> PResult<'_, Expr> {
-    use super::super::stmt::simple::*;
-    if level <= PREC_CONCAT {
-        concat_expr(input)
-    } else {
-        additive_expr(input)
-    }
-}
-
-/// Parse an operand at multiplicative level.
-pub(super) fn multiplicative_operand(input: &str) -> PResult<'_, Expr> {
-    multiplicative_expr(input)
-}
-
-/// Parse an operand at power level.
-pub(super) fn power_operand(input: &str) -> PResult<'_, Expr> {
-    power_expr(input)
-}
-
-fn or_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = assign_or_and_expr(input, mode)?;
-    loop {
-        let (r, _) = ws(rest)?;
-        if let Some((op @ (LogicalOp::Or | LogicalOp::XorXor | LogicalOp::OrElse), len)) =
-            parse_word_logical_op(r)
-        {
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = if mode == ExprMode::Full {
-                assign_or_and_expr(r, mode).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after 'or'/'xor'/'orelse'",
-                        r.len(),
-                    )
-                })?
-            } else {
-                assign_or_and_expr(r, mode)?
-            };
-            left = Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right),
-            };
-            rest = r;
-            continue;
-        }
-        break;
-    }
-    Ok((rest, left))
-}
-
-/// Low-precedence OR/XOR chain for ternary condition parsing.
-/// This intentionally excludes assignment expressions so:
-/// `a = b ?? c !! d` parses as `a = (b ?? c !! d)`.
-fn or_expr_no_assign_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = and_expr_no_assign_mode(input, mode)?;
-    loop {
-        let (r, _) = ws(rest)?;
-        if let Some((op @ (LogicalOp::Or | LogicalOp::XorXor | LogicalOp::OrElse), len)) =
-            parse_word_logical_op(r)
-        {
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = and_expr_no_assign_mode(r, mode)?;
-            left = Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right),
-            };
-            rest = r;
-            continue;
-        }
-        break;
-    }
-    Ok((rest, left))
-}
-
-/// Assignment expressions at the or/and level: $var = expr
-/// This sits between or/and and not in precedence.
-fn assign_or_and_expr(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    and_expr_mode(input, mode)
-}
-
-/// Low-precedence: and / andthen / notandthen
-fn and_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = assign_not_expr_mode(input, mode)?;
-    loop {
-        let (r, _) = ws(rest)?;
-        if let Some((op @ (LogicalOp::And | LogicalOp::AndThen | LogicalOp::NotAndThen), len)) =
-            parse_word_logical_op(r)
-        {
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = if mode == ExprMode::Full {
-                assign_not_expr_mode(r, mode).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after 'and'/'andthen'/'notandthen'",
-                        r.len(),
-                    )
-                })?
-            } else {
-                assign_not_expr_mode(r, mode)?
-            };
-            left = Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right),
-            };
-            rest = r;
-            continue;
-        }
-        break;
-    }
-    Ok((rest, left))
-}
-
-fn and_expr_no_assign_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = not_expr_mode(input, mode)?;
-    loop {
-        let (r, _) = ws(rest)?;
-        if let Some((op @ (LogicalOp::And | LogicalOp::AndThen | LogicalOp::NotAndThen), len)) =
-            parse_word_logical_op(r)
-        {
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = not_expr_mode(r, mode)?;
-            left = Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right),
-            };
-            rest = r;
-            continue;
-        }
-        break;
-    }
-    Ok((rest, left))
-}
-
-fn assign_not_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (rest, expr) = not_expr_mode(input, mode)?;
-    let (r, _) = ws(rest)?;
-
-    // Try compound assignment operators (+=, *=, //=, etc.) before simple =
-    if let Some((after_op, op)) = parse_compound_assign_op(r) {
-        let (after_ws, _) = ws(after_op)?;
-        if let Ok((r, rhs)) = parse_assignment_rhs_mode(after_ws, mode)
-            && let Ok(result) = build_compound_assign_expr(expr.clone(), op, rhs)
-        {
-            return Ok((r, result));
-        }
-    }
-
-    if !(r.starts_with('=') && !r.starts_with("==") && !r.starts_with("=>")) {
-        return Ok((rest, expr));
-    }
-
-    let r = &r[1..];
-    let (r, _) = ws(r)?;
-    // Item assignment (`=`) to a bare scalar variable binds TIGHTER than the
-    // comma operator, so an embedded `$x = 1, 2` parses as `($x = 1), 2`. List
-    // assignment to an `@`/`%` container — and the parenthesized list-lvalue form
-    // `($x) = 1, 2` — keeps the comma-absorbing RHS. We test the pre-unwrap
-    // expression so a grouped `($x)` stays on the list-assignment path. (The
-    // statement-level counterpart lives in `stmt/assign.rs::assign_stmt`.)
-    let scalar_item_assign = matches!(&expr, Expr::Var(_));
-    let (r, rhs) = if scalar_item_assign {
-        ternary_mode(r, mode)?
-    } else {
-        parse_assignment_rhs_mode(r, mode)?
-    };
-
-    let expr = unwrap_grouped_lvalue(expr);
-    match expr {
-        Expr::Var(name) => Ok((
-            r,
-            Expr::AssignExpr {
-                name,
-                expr: Box::new(rhs),
-                is_bind: false,
-            },
-        )),
-        Expr::ArrayVar(name) => Ok((
-            r,
-            Expr::AssignExpr {
-                name: format!("@{}", name),
-                expr: Box::new(rhs),
-                is_bind: false,
-            },
-        )),
-        Expr::HashVar(name) => Ok((
-            r,
-            Expr::AssignExpr {
-                name: format!("%{}", name),
-                expr: Box::new(rhs),
-                is_bind: false,
-            },
-        )),
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-        } => {
-            if let Expr::Call { name, args } = target.as_ref()
-                && name == "__mutsu_subscript_adverb"
-                && args.len() >= 3
-                && matches!(index.as_ref(), Expr::Literal(Value::Int(1)))
-                && matches!(&args[2], Expr::Literal(Value::Str(mode)) if mode.as_str() == "kv" || mode.as_str() == "not-kv")
-            {
-                return Ok((
-                    r,
-                    Expr::IndexAssign {
-                        target: Box::new(args[0].clone()),
-                        index: Box::new(args[1].clone()),
-                        value: Box::new(rhs),
-                        is_positional,
-                    },
-                ));
-            }
-            Ok((
-                r,
-                Expr::IndexAssign {
-                    target,
-                    index,
-                    value: Box::new(rhs),
-                    is_positional,
-                },
-            ))
-        }
-        Expr::Call { name, args } => Ok((
-            r,
-            Expr::Call {
-                name: Symbol::intern("__mutsu_assign_named_sub_lvalue"),
-                args: vec![
-                    Expr::Literal(Value::str(name.resolve())),
-                    Expr::ArrayLiteral(args),
-                    rhs,
-                ],
-            },
-        )),
-        Expr::MultiDimIndex { target, dimensions } => Ok((
-            r,
-            Expr::MultiDimIndexAssign {
-                target,
-                dimensions,
-                value: Box::new(rhs),
-            },
-        )),
-        Expr::BareWord(name) => Ok((
-            r,
-            Expr::AssignExpr {
-                name,
-                expr: Box::new(rhs),
-                is_bind: false,
-            },
-        )),
-        Expr::CallOn { target, args } => Ok((
-            r,
-            if args.is_empty() {
-                if let Expr::ArrayLiteral(items) = *target.clone() {
-                    if let Some(expr) = list_lvalue_assign_expr(items, rhs.clone()) {
-                        expr
-                    } else {
-                        Expr::Call {
-                            name: Symbol::intern("__mutsu_assign_callable_lvalue"),
-                            args: vec![*target, Expr::ArrayLiteral(args), rhs],
-                        }
-                    }
-                } else {
-                    Expr::Call {
-                        name: Symbol::intern("__mutsu_assign_callable_lvalue"),
-                        args: vec![*target, Expr::ArrayLiteral(args), rhs],
-                    }
-                }
-            } else {
-                Expr::Call {
-                    name: Symbol::intern("__mutsu_assign_callable_lvalue"),
-                    args: vec![*target, Expr::ArrayLiteral(args), rhs],
-                }
-            },
-        )),
-        _ => Ok((rest, expr)),
-    }
-}
-
-fn assignment_ro_expr(lhs: Expr, rhs: Expr) -> Expr {
-    Expr::DoBlock {
-        body: vec![
-            Stmt::Expr(lhs),
-            Stmt::Expr(rhs),
-            Stmt::Expr(Expr::Call {
-                name: Symbol::intern("__mutsu_assignment_ro"),
-                args: Vec::new(),
-            }),
-        ],
-        label: None,
-    }
-}
-
-fn unwrap_grouped_lvalue(target: Expr) -> Expr {
-    match target {
-        Expr::Grouped(inner) => unwrap_grouped_lvalue(*inner),
-        other => other,
-    }
-}
-
-fn assign_to_target_expr(target: Expr, value: Expr) -> Expr {
-    match unwrap_grouped_lvalue(target) {
-        Expr::Var(name) => Expr::AssignExpr {
-            name,
-            expr: Box::new(value),
-            is_bind: false,
-        },
-        Expr::ArrayVar(name) => Expr::AssignExpr {
-            name: format!("@{}", name),
-            expr: Box::new(value),
-            is_bind: false,
-        },
-        Expr::HashVar(name) => Expr::AssignExpr {
-            name: format!("%{}", name),
-            expr: Box::new(value),
-            is_bind: false,
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-        } => Expr::IndexAssign {
-            target,
-            index,
-            value: Box::new(value),
-            is_positional,
-        },
-        Expr::MultiDimIndex { target, dimensions } => Expr::MultiDimIndexAssign {
-            target,
-            dimensions,
-            value: Box::new(value),
-        },
-        Expr::Call { name, args } => Expr::Call {
-            name: Symbol::intern("__mutsu_assign_named_sub_lvalue"),
-            args: vec![
-                Expr::Literal(Value::str(name.resolve())),
-                Expr::ArrayLiteral(args),
-                value,
-            ],
-        },
-        Expr::CallOn { target, args } => Expr::Call {
-            name: Symbol::intern("__mutsu_assign_callable_lvalue"),
-            args: vec![*target, Expr::ArrayLiteral(args), value],
-        },
-        Expr::SymbolicDeref { sigil, expr } => Expr::SymbolicDerefAssign {
-            sigil,
-            expr,
-            value: Box::new(value),
-        },
-        Expr::IndirectTypeLookup(inner) => Expr::IndirectTypeLookupAssign {
-            expr: inner,
-            value: Box::new(value),
-        },
-        Expr::DoStmt(stmt) => {
-            if let Stmt::VarDecl {
-                name,
-                type_constraint,
-                is_state,
-                is_our,
-                is_dynamic,
-                is_export,
-                export_tags,
-                custom_traits,
-                where_constraint,
-                ..
-            } = *stmt
-            {
-                Expr::DoStmt(Box::new(Stmt::VarDecl {
-                    name,
-                    expr: value,
-                    type_constraint,
-                    is_state,
-                    is_our,
-                    is_dynamic,
-                    is_export,
-                    export_tags,
-                    custom_traits,
-                    where_constraint,
-                }))
-            } else {
-                assignment_ro_expr(Expr::DoStmt(stmt), value)
-            }
-        }
-        other => assignment_ro_expr(other, value),
-    }
-}
-
-fn build_compound_assign_target_expr(target: Expr, op_name: &str, value: Expr) -> Expr {
-    let target = unwrap_grouped_lvalue(target);
-    if op_name == "=" {
-        return assign_to_target_expr(target, value);
-    }
-    let Some(op) = compound_assign_op_from_name(op_name) else {
-        return assignment_ro_expr(target, value);
-    };
-    // Bare `* op= value` curries to a WhateverCode that mutates its topic,
-    // i.e. `{ $_ op= value }`. Used for e.g. `@a.map(* *= 2)`.
-    if matches!(target, Expr::Whatever) {
-        let body = build_compound_assign_target_expr(Expr::Var("_".to_string()), op_name, value);
-        return Expr::Lambda {
-            param: "_".to_string(),
-            body: vec![crate::ast::Stmt::Expr(body)],
-            is_whatever_code: true,
-        };
-    }
-    match target {
-        Expr::Var(name) => Expr::AssignExpr {
-            name: name.clone(),
-            expr: Box::new(compound_assigned_value_expr(Expr::Var(name), op, value)),
-            is_bind: false,
-        },
-        Expr::ArrayVar(name) => Expr::AssignExpr {
-            name: format!("@{}", name.clone()),
-            expr: Box::new(compound_assigned_value_expr(
-                Expr::ArrayVar(name),
-                op,
-                value,
-            )),
-            is_bind: false,
-        },
-        Expr::HashVar(name) => Expr::AssignExpr {
-            name: format!("%{}", name.clone()),
-            expr: Box::new(compound_assigned_value_expr(Expr::HashVar(name), op, value)),
-            is_bind: false,
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-            ..
-        } => {
-            let lhs_expr = Expr::Index {
-                target: target.clone(),
-                index: index.clone(),
-                is_positional,
-            };
-            Expr::IndexAssign {
-                target,
-                index,
-                value: Box::new(compound_assigned_value_expr(lhs_expr, op, value)),
-                is_positional: true,
-            }
-        }
-        Expr::MultiDimIndex { target, dimensions } => {
-            let lhs_expr = Expr::MultiDimIndex {
-                target: target.clone(),
-                dimensions: dimensions.clone(),
-            };
-            Expr::MultiDimIndexAssign {
-                target,
-                dimensions,
-                value: Box::new(compound_assigned_value_expr(lhs_expr, op, value)),
-            }
-        }
-        Expr::AssignExpr {
-            name,
-            expr,
-            is_bind: _,
-        } => {
-            if op_name == "=" {
-                return Expr::AssignExpr {
-                    name,
-                    expr: Box::new(value),
-                    is_bind: false,
-                };
-            }
-            let Some(op) = compound_assign_op_from_name(op_name) else {
-                return assignment_ro_expr(
-                    Expr::AssignExpr {
-                        name,
-                        expr,
-                        is_bind: false,
-                    },
-                    value,
-                );
-            };
-            Expr::AssignExpr {
-                name: name.clone(),
-                expr: Box::new(compound_assigned_value_expr(
-                    Expr::AssignExpr {
-                        name,
-                        expr,
-                        is_bind: false,
-                    },
-                    op,
-                    value,
-                )),
-                is_bind: false,
-            }
-        }
-        Expr::MethodCall {
-            target,
-            name,
-            args,
-            modifier: _,
-            quoted: _,
-        } if name == "AT-POS" && args.len() == 1 => {
-            let index = args.into_iter().next().unwrap_or(Expr::Literal(Value::Nil));
-            build_compound_assign_target_expr(
-                Expr::Index {
-                    target,
-                    index: Box::new(index),
-                    is_positional: true,
-                },
-                op_name,
-                value,
-            )
-        }
-        other => assignment_ro_expr(other, value),
-    }
-}
-
-fn list_lvalue_assign_expr(items: Vec<Expr>, rhs: Expr) -> Option<Expr> {
-    let mut saw_whatever = false;
-    let mut lvalues: Vec<Expr> = Vec::new();
-    for item in items {
-        if matches!(item, Expr::Whatever) {
-            saw_whatever = true;
-            continue;
-        }
-        lvalues.push(item);
-    }
-    if !saw_whatever || lvalues.len() != 1 {
-        return None;
-    }
-    match lvalues.into_iter().next()? {
-        Expr::Var(name) => Some(Expr::AssignExpr {
-            name,
-            expr: Box::new(rhs),
-            is_bind: false,
-        }),
-        Expr::ArrayVar(name) => Some(Expr::AssignExpr {
-            name: format!("@{}", name),
-            expr: Box::new(rhs),
-            is_bind: false,
-        }),
-        Expr::HashVar(name) => Some(Expr::AssignExpr {
-            name: format!("%{}", name),
-            expr: Box::new(rhs),
-            is_bind: false,
-        }),
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-        } => Some(Expr::IndexAssign {
-            target,
-            index,
-            value: Box::new(rhs),
-            is_positional,
-        }),
-        _ => None,
-    }
-}
-
-fn parse_assignment_rhs_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (rest, first) = ternary_mode(input, mode)?;
-    let (r, _) = ws(rest)?;
-    if !r.starts_with(',') || r.starts_with(",,") {
-        return Ok((rest, first));
-    }
-
-    let mut items = vec![first];
-    let mut cursor = r;
-    loop {
-        let (r2, _) = parse_char(cursor, ',')?;
-        let (r2, _) = ws(r2)?;
-        if r2.is_empty() || r2.starts_with(';') || r2.starts_with('}') || r2.starts_with(')') {
-            return Ok((r2, Expr::ArrayLiteral(items)));
-        }
-        let (r3, next) = ternary_mode(r2, mode)?;
-        items.push(next);
-        let (r3, _) = ws(r3)?;
-        if !r3.starts_with(',') || r3.starts_with(",,") {
-            return Ok((r3, Expr::ArrayLiteral(items)));
-        }
-        cursor = r3;
-    }
-}
-
-fn not_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    if input.starts_with("not")
-        && !is_ident_char(input.as_bytes().get(3).copied())
-        && !input[3..].starts_with('(')
-        && !{
-            let trimmed = input[3..].trim_start();
-            trimmed.starts_with("=>") && !trimmed.starts_with("==>")
-        }
-    {
-        let r = &input[3..];
-        let (r, _) = ws(r)?;
-        // Use assign_not_expr_mode so that `not $x = 42` parses as `not($x = 42)`
-        // since item assignment is tighter than loose unary not/so.
-        let (r, expr) = assign_not_expr_mode(r, mode)?;
-        return Ok((
-            r,
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr: Box::new(expr),
-            },
-        ));
-    }
-    if input.starts_with("so")
-        && !is_ident_char(input.as_bytes().get(2).copied())
-        && !input[2..].starts_with('(')
-    {
-        let r = &input[2..];
-        let (r, _) = ws(r)?;
-        // Use assign_not_expr_mode so that `so $x = 42` parses as `so($x = 42)`
-        // since item assignment is tighter than loose unary not/so.
-        let (r, expr) = assign_not_expr_mode(r, mode)?;
-        return Ok((
-            r,
-            Expr::Unary {
-                op: TokenKind::Question,
-                expr: Box::new(expr),
-            },
-        ));
-    }
-    flipflop_expr_mode(input, mode)
-}
-
-/// Flip-flop operators (`ff`/`fff` and the endpoint-excluding forms) at
-/// Raku's *conditional* precedence (`?? !!` level): looser than tight-or
-/// (`||`), tight-and (`&&`), and the chaining comparisons (`==` etc.), but
-/// tighter than item assignment (`=`) and the comma/list-infix operators. So
-/// `$n == 3 ff $n == 6` parses as `($n == 3) ff ($n == 6)`.
-fn flipflop_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = or_or_expr_mode(input, mode)?;
-    let mut current_assoc_key: Option<String> = None;
-    loop {
-        let (r, _) = ws(rest)?;
-        let Some((name, len)) = parse_flipflop_infix(r) else {
-            break;
-        };
-        if let Some(prev) = current_assoc_key.as_deref()
-            && prev != name
-        {
-            return Err(non_list_associative_error(prev, &name));
-        }
-        let r = &r[len..];
-        let (r, _) = ws(r)?;
-        let (r, right) = or_or_expr_mode(r, mode).map_err(|err| {
-            enrich_expected_error(err, "expected expression after flip-flop operator", r.len())
-        })?;
-        left = Expr::InfixFunc {
-            name: name.clone(),
-            left: Box::new(left),
-            right: vec![right],
-            modifier: None,
-        };
-        current_assoc_key = Some(name);
-        rest = r;
-    }
-    Ok((rest, left))
-}
-
-/// || , ^^ , and //
-fn or_or_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = and_and_expr_mode(input, mode)?;
-    let mut last_list_assoc_op: Option<LogicalOp> = None;
-    loop {
-        let (r, _) = ws(rest)?;
-        // Check for negated logical ops first: !|| , !^^
-        let (negated, op_result) = if let Some((op, len)) = parse_negated_logical_op(r) {
-            if matches!(op, LogicalOp::OrOr | LogicalOp::XorXor) {
-                (true, Some((op, len)))
-            } else {
-                (false, None)
-            }
-        } else {
-            (false, None)
-        };
-        let (op, len) = if let Some(pair) = op_result {
-            pair
-        } else if let Some(pair) = parse_or_or_op(r) {
-            pair
-        } else {
-            break;
-        };
-        if matches!(op, LogicalOp::Min | LogicalOp::Max) {
-            if let Some(prev) = last_list_assoc_op
-                && prev != op
-            {
-                {
-                    let lhs = match prev {
-                        LogicalOp::Min => "min",
-                        LogicalOp::Max => "max",
-                        _ => unreachable!(),
-                    };
-                    let rhs = match op {
-                        LogicalOp::Min => "min",
-                        LogicalOp::Max => "max",
-                        _ => unreachable!(),
-                    };
-                    return Err(syntax_exception(
-                        "X::Syntax::NonAssociative",
-                        format!(
-                            "Only identical operators may be list associative; since '{}' and '{}' differ, they are non-associative and you need to clarify with parentheses",
-                            lhs, rhs
-                        ),
-                    ));
-                }
-            }
-            last_list_assoc_op = Some(op);
-        }
-        let r = &r[len..];
-        let (r, _) = ws(r)?;
-        let (r, right) = if mode == ExprMode::Full {
-            if is_loose_not_or_so_prefix(r) {
-                not_expr_mode(r, mode).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after logical operator",
-                        r.len(),
-                    )
-                })?
-            } else {
-                and_and_expr_mode(r, mode).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after logical operator",
-                        r.len(),
-                    )
-                })?
-            }
-        } else if is_loose_not_or_so_prefix(r) {
-            not_expr_mode(r, mode)?
-        } else {
-            and_and_expr_mode(r, mode)?
-        };
-        let binary = Expr::Binary {
-            left: Box::new(left),
-            op: op.token_kind(),
-            right: Box::new(right),
-        };
-        left = if negated {
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr: Box::new(binary),
-            }
-        } else {
-            binary
-        };
-        rest = r;
-    }
-    Ok((rest, left))
-}
-
-fn is_loose_not_or_so_prefix(input: &str) -> bool {
-    (input.starts_with("not")
-        && !is_ident_char(input.as_bytes().get(3).copied())
-        && !input[3..].starts_with('('))
-        || (input.starts_with("so")
-            && !is_ident_char(input.as_bytes().get(2).copied())
-            && !input[2..].starts_with('('))
-}
-
-/// &&
-fn and_and_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = comparison_expr_mode(input, mode)?;
-    loop {
-        let (r, _) = ws(rest)?;
-        // Check for negated !&&
-        let (negated, op, len) = if let Some((op, len)) = parse_negated_logical_op(r) {
-            if matches!(op, LogicalOp::AndAnd) {
-                (true, op, len)
-            } else if let Some((op, len)) = parse_and_and_op(r) {
-                (false, op, len)
-            } else {
-                break;
-            }
-        } else if let Some((op, len)) = parse_and_and_op(r) {
-            (false, op, len)
-        } else {
-            break;
-        };
-        let r = &r[len..];
-        let (r, _) = ws(r)?;
-        // `A && not B` / `A && so B`: loose unary `not`/`so` on the RHS of a
-        // tight operator binds LOOSELY (`A && (not (B ...))`), so delegate to the
-        // loose-unary parser instead of the tighter comparison operand.
-        let (r, right) = if is_loose_not_or_so_prefix(r) {
-            not_expr_mode(r, mode)?
-        } else if mode == ExprMode::Full {
-            comparison_expr_mode(r, mode).map_err(|err| {
-                enrich_expected_error(err, "expected expression after '&&'", r.len())
-            })?
-        } else {
-            comparison_expr_mode(r, mode)?
-        };
-        let binary = Expr::Binary {
-            left: Box::new(left),
-            op: op.token_kind(),
-            right: Box::new(right),
-        };
-        left = if negated {
-            Expr::Unary {
-                op: TokenKind::Bang,
-                expr: Box::new(binary),
-            }
-        } else {
-            binary
-        };
-        rest = r;
-    }
-    Ok((rest, left))
-}
-
-/// Boolean bitwise / junction: ?| ?& ?^ | & ^
-fn junctive_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    let next_fn: fn(&str) -> PResult<'_, Expr> = match mode {
-        ExprMode::Full => sequence_expr,
-        ExprMode::NoSequence => list_infix_expr,
-        ExprMode::NoSequenceNoFeed => range_expr,
-    };
-    let (mut rest, mut left) = next_fn(input)?;
-    let mut last_junction: Option<JunctionInfixOp> = None;
-    loop {
-        let (r, _) = ws(rest)?;
-        if let Some((op, len)) = parse_junctive_op(r) {
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = if mode == ExprMode::Full {
-                next_fn(r).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after junctive operator",
-                        r.len(),
-                    )
-                })?
-            } else {
-                next_fn(r)?
-            };
-            left = Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right),
-            };
-            rest = r;
-            continue;
-        }
-        // Junction infix operators: |, &, ^
-        if let Some((op, len)) = parse_junction_infix_op(r) {
-            if let Some(prev) = last_junction
-                && matches!(
-                    (prev, op),
-                    (JunctionInfixOp::Any, JunctionInfixOp::One)
-                        | (JunctionInfixOp::One, JunctionInfixOp::Any)
-                )
-            {
-                return Err(non_list_associative_error(
-                    match prev {
-                        JunctionInfixOp::Any => "|",
-                        JunctionInfixOp::All => "&",
-                        JunctionInfixOp::One => "^",
-                    },
-                    match op {
-                        JunctionInfixOp::Any => "|",
-                        JunctionInfixOp::All => "&",
-                        JunctionInfixOp::One => "^",
-                    },
-                ));
-            }
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = if mode == ExprMode::Full {
-                next_fn(r).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after junction operator",
-                        r.len(),
-                    )
-                })?
-            } else {
-                next_fn(r)?
-            };
-            left = Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right),
-            };
-            last_junction = Some(op);
-            rest = r;
-            continue;
-        }
-        break;
-    }
-    Ok((rest, left))
-}
-
-/// List-infix operators only: Z, X, meta-ops, infix funcs.
-/// Used in NoSequence mode (e.g. inside parenthesized expressions).
-fn list_infix_expr(input: &str) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = range_expr(input)?;
-    let mut current_assoc_key = None;
-    rest = parse_list_infix_loop(rest, input, &mut left, &mut current_assoc_key)?;
-    Ok((rest, left))
-}
-
-/// Sequence (..., ...^) and list-infix (Z, X, meta-ops, infix funcs).
-/// All have Raku precedence level f= (list associative).
-fn sequence_expr(input: &str) -> PResult<'_, Expr> {
-    let (mut rest, mut left) = range_expr(input)?;
-    let mut current_assoc_key: Option<String> = None;
-
-    // Helper: wrap LHS WhateverCode before building the sequence node.
-    fn maybe_wrap_lhs(left: &mut Expr) {
-        if contains_whatever(left) && !matches!(left, Expr::Whatever) {
-            *left = wrap_whatevercode(left);
-        }
-    }
-
-    loop {
-        let (r, _) = ws(rest)?;
-
-        // ...^ / …^ (exclusive-end sequence), ... / … (sequence)
-        if let Some((r2, op, op_str)) = strip_sequence_op(r) {
-            if let Some(prev) = current_assoc_key.as_deref()
-                && prev != op_str
-            {
-                return Err(non_list_associative_error(prev, op_str));
-            }
-            let (r2, _) = ws(r2)?;
-            let (r2, mut right) =
-                comparison_expr_mode(r2, ExprMode::NoSequence).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        format!("expected expression after '{op_str}'").as_str(),
-                        r2.len(),
-                    )
-                })?;
-            if contains_whatever(&right) && !matches!(right, Expr::Whatever) {
-                right = wrap_whatevercode(&right);
-            }
-            maybe_wrap_lhs(&mut left);
-            left = Expr::Binary {
-                left: Box::new(left),
-                op,
-                right: Box::new(right),
-            };
-            current_assoc_key = Some(op_str.to_string());
-            rest = r2;
-            continue;
-        }
-        // Try Z/X/meta/infix-func operators
-        let new_rest = parse_list_infix_loop(rest, input, &mut left, &mut current_assoc_key)?;
-        if new_rest.len() < rest.len() {
-            rest = new_rest;
-            continue;
-        }
-        break;
-    }
-    Ok((rest, left))
-}
-
-/// Shared loop for Z/X meta operators and infix function calls.
-/// Modifies `left` in place and returns the remaining input.
-fn parse_list_infix_loop<'a>(
-    input: &'a str,
-    orig_input: &str,
-    left: &mut Expr,
-    current_assoc_key: &mut Option<String>,
-) -> Result<&'a str, PError> {
-    let mut rest = input;
-    loop {
-        let (r, _) = ws(rest)?;
-        if r.starts_with("X.") && !r.starts_with("X..") {
-            let after = &r[2..];
-            let after_trimmed = after.trim_start();
-            if after_trimmed.starts_with('"') || after_trimmed.starts_with('\'') {
-                return Err(PError::expected_at(
-                    "X::Obsolete: Perl . is dead. Please use ~ to concatenate strings.",
-                    r,
-                ));
-            }
-            return Err(PError::expected_at(
-                "X::Syntax::CannotMeta: Cannot do . because it is too fiddly",
-                r,
-            ));
-        }
-        if r.starts_with("Z.") && !r.starts_with("Z..") {
-            let after = &r[2..];
-            let after_trimmed = after.trim_start();
-            if after_trimmed.starts_with('"') || after_trimmed.starts_with('\'') {
-                return Err(PError::expected_at(
-                    "X::Obsolete: Perl . is dead. Please use ~ to concatenate strings.",
-                    r,
-                ));
-            }
-            return Err(PError::expected_at(
-                "X::Syntax::CannotMeta: Cannot do . because it is too fiddly",
-                r,
-            ));
-        }
-        // Infixed function call: [&func], R[&func], X[&func], Z[&func]
-        if let Some((modifier, name, len)) = parse_infix_func_op(r) {
-            let op_key = if let Some(modifier) = modifier.as_deref() {
-                format!("{modifier}[{name}]")
-            } else {
-                format!("[{name}]")
-            };
-            if let Some(prev) = current_assoc_key.as_deref()
-                && prev != op_key
-            {
-                return Err(non_list_associative_error(prev, &op_key));
-            }
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, mut right_exprs) = if modifier.as_deref() == Some("X") {
-                parse_comma_list_of_range(r)?
-            } else {
-                let (r, expr) = range_expr(r).map_err(|err| {
-                    enrich_expected_error(
-                        err,
-                        "expected expression after infixed function",
-                        r.len(),
-                    )
-                })?;
-                (r, vec![expr])
-            };
-            // Collect trailing colonpair adverbs (e.g., `3 zin 4 :x(5)`)
-            let mut r = r;
-            loop {
-                let (r2, _) = ws(r)?;
-                if r2.starts_with(':')
-                    && !r2.starts_with("::")
-                    && let Ok((r3, adverb)) = super::super::primary::colonpair_expr(r2)
-                {
-                    right_exprs.push(adverb);
-                    r = r3;
-                } else {
-                    break;
-                }
-            }
-            *left = Expr::InfixFunc {
-                name,
-                left: Box::new(left.clone()),
-                right: right_exprs,
-                modifier,
-            };
-            *current_assoc_key = Some(op_key);
-            rest = r;
-            continue;
-        }
-        // Negated bracket infix operators: ![op]
-        if let Some(r_after_bang) = r.strip_prefix('!')
-            && let Some(bracket_infix) = parse_bracket_infix_op(r_after_bang)
-        {
-            match bracket_infix {
-                BracketInfix::PlainOp(op, len) => {
-                    let r = &r[1 + len..];
-                    let (r, _) = ws(r)?;
-                    let (r, right) = range_expr(r).map_err(|err| {
-                        enrich_expected_error(
-                            err,
-                            "expected expression after negated bracket infix op",
-                            r.len(),
-                        )
-                    })?;
-                    let binary = if let Some(tk) = op_str_to_token_kind(&op) {
-                        Expr::Binary {
-                            left: Box::new(left.clone()),
-                            op: tk,
-                            right: Box::new(right),
-                        }
-                    } else {
-                        Expr::InfixFunc {
-                            name: op,
-                            left: Box::new(left.clone()),
-                            right: vec![right],
-                            modifier: None,
-                        }
-                    };
-                    *left = Expr::Unary {
-                        op: TokenKind::Bang,
-                        expr: Box::new(binary),
-                    };
-                    rest = r;
-                    continue;
-                }
-                BracketInfix::MetaOp(meta, op, len) => {
-                    let r = &r[1 + len..];
-                    let (r, _) = ws(r)?;
-                    let (r, right) = range_expr(r).map_err(|err| {
-                        enrich_expected_error(
-                            err,
-                            "expected expression after negated bracket meta infix op",
-                            r.len(),
-                        )
-                    })?;
-                    *left = Expr::Unary {
-                        op: TokenKind::Bang,
-                        expr: Box::new(Expr::MetaOp {
-                            meta,
-                            op,
-                            left: Box::new(left.clone()),
-                            right: Box::new(right),
-                        }),
-                    };
-                    rest = r;
-                    continue;
-                }
-                BracketInfix::UserInfix(name, len) => {
-                    let r = &r[1 + len..];
-                    let (r, _) = ws(r)?;
-                    let (r, right) = range_expr(r).map_err(|err| {
-                        enrich_expected_error(
-                            err,
-                            "expected expression after negated bracket user infix op",
-                            r.len(),
-                        )
-                    })?;
-                    *left = Expr::Unary {
-                        op: TokenKind::Bang,
-                        expr: Box::new(Expr::InfixFunc {
-                            name,
-                            left: Box::new(left.clone()),
-                            right: vec![right],
-                            modifier: None,
-                        }),
-                    };
-                    rest = r;
-                    continue;
-                }
-            }
-        }
-        // Bracket infix operators: [+], [R-], [Z*], [Z[cmp]], [blue], etc.
-        if let Some(bracket_infix) = parse_bracket_infix_op(r) {
-            match bracket_infix {
-                BracketInfix::PlainOp(op, len) => {
-                    let r = &r[len..];
-                    let (r, _) = ws(r)?;
-                    let (r, right) = range_expr(r).map_err(|err| {
-                        enrich_expected_error(
-                            err,
-                            "expected expression after bracket infix op",
-                            r.len(),
-                        )
-                    })?;
-                    // Resolve plain op to a Binary expr
-                    if let Some(tk) = op_str_to_token_kind(&op) {
-                        *left = Expr::Binary {
-                            left: Box::new(left.clone()),
-                            op: tk,
-                            right: Box::new(right),
-                        };
-                    } else {
-                        // Unknown plain op: treat as user-defined infix
-                        *left = Expr::InfixFunc {
-                            name: op,
-                            left: Box::new(left.clone()),
-                            right: vec![right],
-                            modifier: None,
-                        };
-                    }
-                    rest = r;
-                    continue;
-                }
-                BracketInfix::MetaOp(meta, op, len) => {
-                    let r = &r[len..];
-                    let (r, _) = ws(r)?;
-                    let needs_comma_list = meta == "Z"
-                        || meta == "X"
-                        || op == "..."
-                        || op == "...^"
-                        || op == "…"
-                        || op == "…^";
-                    let (r, right) = if needs_comma_list {
-                        let (r, items) = parse_comma_list_of_range_raw(r)?;
-                        if items.len() == 1 {
-                            (r, items.into_iter().next().unwrap())
-                        } else {
-                            (r, Expr::ArrayLiteral(items))
-                        }
-                    } else {
-                        range_expr(r).map_err(|err| {
-                            enrich_expected_error(
-                                err,
-                                "expected expression after bracket meta infix op",
-                                r.len(),
-                            )
-                        })?
-                    };
-                    let (r, right) = if meta == "X" {
-                        let (r_ws, _) = ws(r)?;
-                        if let Some(after_op) = r_ws.strip_prefix("~~") {
-                            let (after_op, _) = ws(after_op)?;
-                            let (r_after_rhs, rhs_expr) =
-                                junctive_expr_mode(after_op, ExprMode::Full).map_err(|err| {
-                                    enrich_expected_error(
-                                        err,
-                                        "expected expression after smartmatch",
-                                        after_op.len(),
-                                    )
-                                })?;
-                            (
-                                r_after_rhs,
-                                Expr::Binary {
-                                    left: Box::new(right),
-                                    op: TokenKind::SmartMatch,
-                                    right: Box::new(rhs_expr),
-                                },
-                            )
-                        } else {
-                            (r_ws, right)
-                        }
-                    } else {
-                        (r, right)
-                    };
-                    *left = Expr::MetaOp {
-                        meta,
-                        op,
-                        left: Box::new(left.clone()),
-                        right: Box::new(right),
-                    };
-                    rest = r;
-                    continue;
-                }
-                BracketInfix::UserInfix(name, len) => {
-                    let r = &r[len..];
-                    let (r, _) = ws(r)?;
-                    let (r, right) = range_expr(r).map_err(|err| {
-                        enrich_expected_error(
-                            err,
-                            "expected expression after bracket user infix op",
-                            r.len(),
-                        )
-                    })?;
-                    *left = Expr::InfixFunc {
-                        name,
-                        left: Box::new(left.clone()),
-                        right: vec![right],
-                        modifier: None,
-                    };
-                    rest = r;
-                    continue;
-                }
-            }
-        }
-        // Meta operators: R-, X+, Z~, bare Z, bare X, R[...], etc.
-        if let Some((stripped, meta, op_name)) = parse_meta_compound_assign_op(r)
-            && meta == "R"
-        {
-            let (r, _) = ws(stripped)?;
-            let (r, rhs) = parse_assign_expr_or_comma(r).map_err(|err| {
-                enrich_expected_error(
-                    err,
-                    "expected expression after reverse meta compound assignment",
-                    r.len(),
-                )
-            })?;
-            *left = build_compound_assign_target_expr(rhs, &op_name, left.clone());
-            rest = r;
-            continue;
-        }
-        if let Some((meta, op, len)) = parse_meta_op(r) {
-            // A reversed range meta-op (`R..`, `R^..`, `R..^`, `R^..^`) carries the
-            // same precedence worry as a plain range: `|4 R.. 5` / `~4 R.. 5` mean
-            // `|(4 R.. 5)` / `~(4 R.. 5)`, not `(|4) R.. 5`. Fire the worry off the
-            // ORIGINAL input text (parentheses are transparent in the AST), matching
-            // the direct-range path in `range_expr`.
-            if matches!(op.as_str(), ".." | "..^" | "^.." | "^..^") {
-                check_range_precedence_worry(orig_input)?;
-            }
-            let op_key = format!("{meta}{op}");
-            if let Some(prev) = current_assoc_key.as_deref()
-                && prev != op_key
-            {
-                return Err(non_list_associative_error(prev, &op_key));
-            }
-            if meta == "X"
-                && op == "cmp"
-                && matches!(
-                    &left,
-                    Expr::MetaOp {
-                        meta: prev_meta,
-                        op: prev_op,
-                        ..
-                    } if prev_meta == "X" && prev_op == "cmp"
-                )
-            {
-                return Err(non_associative_error("Xcmp"));
-            }
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let needs_comma_list = (meta == "Z" || meta == "X")
-                || op == "..."
-                || op == "...^"
-                || op == "…"
-                || op == "…^";
-            let (r, right) = if needs_comma_list {
-                let (r, items) = parse_comma_list_of_range_raw(r)?;
-                if items.len() == 1 {
-                    (r, items.into_iter().next().unwrap())
-                } else {
-                    (r, Expr::ArrayLiteral(items))
-                }
-            } else {
-                range_expr(r).map_err(|err| {
-                    enrich_expected_error(err, "expected expression after meta operator", r.len())
-                })?
-            };
-            let (r, right) = if meta == "X" {
-                let (r_ws, _) = ws(r)?;
-                if let Some(after_op) = r_ws.strip_prefix("~~") {
-                    let (after_op, _) = ws(after_op)?;
-                    let (r_after_rhs, rhs_expr) = junctive_expr_mode(after_op, ExprMode::Full)
-                        .map_err(|err| {
-                            enrich_expected_error(
-                                err,
-                                "expected expression after smartmatch",
-                                after_op.len(),
-                            )
-                        })?;
-                    (
-                        r_after_rhs,
-                        Expr::Binary {
-                            left: Box::new(right),
-                            op: TokenKind::SmartMatch,
-                            right: Box::new(rhs_expr),
-                        },
-                    )
-                } else {
-                    (r_ws, right)
-                }
-            } else {
-                (r, right)
-            };
-            *left = Expr::MetaOp {
-                meta,
-                op,
-                left: Box::new(left.clone()),
-                right: Box::new(right),
-            };
-            *current_assoc_key = Some(op_key);
-            rest = r;
-            continue;
-        }
-        // User-defined infix words (typically via my &infix:<...> = ...),
-        // e.g. `42 same-in-Int "42"`.
-        // Do not span statement boundaries across newlines.
-        let ws_before = &rest[..rest.len() - r.len()];
-        if !ws_before.contains('\n')
-            && let Some((feed_op, len)) = parse_feed_op(r)
-        {
-            let r = &r[len..];
-            let (r, _) = ws(r)?;
-            let (r, right) = (if matches!(feed_op, FeedOp::ToLeft | FeedOp::AppendLeft) {
-                sequence_expr(r)
-            } else {
-                range_expr(r)
-            })
-            .map_err(|err| {
-                enrich_expected_error(err, "expected expression after feed operator", r.len())
-            })?;
-            *left = make_feed_node(feed_op, left.clone(), right);
-            rest = r;
-            continue;
-        }
-        if !ws_before.contains('\n')
-            && let Some((name, len)) = parse_custom_infix_word(r)
-            && super::super::stmt::simple::lookup_custom_infix_precedence(&name)
-                .is_none_or(|level| level <= super::super::stmt::simple::PREC_SEQUENCE)
-        {
-            let mut r = &r[len..];
-            let (r2, _) = ws(r)?;
-            let (r2, right) = range_expr(r2).map_err(|err| {
-                enrich_expected_error(err, "expected expression after infix operator", r.len())
-            })?;
-            let assoc = super::super::stmt::simple::lookup_user_infix_assoc(&name)
-                .unwrap_or_else(|| "left".to_string());
-            let mut args = vec![left.clone(), right];
-            r = r2;
-            if assoc != "left" {
-                loop {
-                    let (r_ws, _) = ws(r)?;
-                    let ws_between = &r[..r.len() - r_ws.len()];
-                    if ws_between.contains('\n') {
-                        break;
-                    }
-                    let Some((next_name, next_len)) = parse_custom_infix_word(r_ws) else {
-                        break;
-                    };
-                    if next_name != name {
-                        break;
-                    }
-                    let r_after_op = &r_ws[next_len..];
-                    let (r_after_op, _) = ws(r_after_op)?;
-                    let (r_after_arg, arg) = range_expr(r_after_op).map_err(|err| {
-                        enrich_expected_error(
-                            err,
-                            "expected expression after infix operator",
-                            r_after_op.len(),
-                        )
-                    })?;
-                    args.push(arg);
-                    r = r_after_arg;
-                }
-            }
-            // Collect trailing colonpair adverbs (e.g., `3 zin 4 :x(5)`)
-            loop {
-                let (r_ws, _) = ws(r)?;
-                if r_ws.starts_with(':')
-                    && !r_ws.starts_with("::")
-                    && let Ok((r3, adverb)) = super::super::primary::colonpair_expr(r_ws)
-                {
-                    args.push(adverb);
-                    r = r3;
-                } else {
-                    break;
-                }
-            }
-            *left = match assoc.as_str() {
-                "right" => {
-                    let mut iter = args.into_iter().rev();
-                    let mut acc = iter
-                        .next()
-                        .unwrap_or(Expr::Literal(crate::value::Value::Nil));
-                    for lhs in iter {
-                        acc = Expr::InfixFunc {
-                            name: name.clone(),
-                            left: Box::new(lhs),
-                            right: vec![acc],
-                            modifier: None,
-                        };
-                    }
-                    acc
-                }
-                "list" | "chain" => Expr::InfixFunc {
-                    name: name.clone(),
-                    left: Box::new(args[0].clone()),
-                    right: args[1..].to_vec(),
-                    modifier: None,
-                },
-                "non" => {
-                    if args.len() > 2 {
-                        return Err(non_associative_error(&name));
-                    }
-                    Expr::InfixFunc {
-                        name: name.clone(),
-                        left: Box::new(args[0].clone()),
-                        right: args[1..].to_vec(),
-                        modifier: None,
-                    }
-                }
-                _ => Expr::InfixFunc {
-                    name: name.clone(),
-                    left: Box::new(args[0].clone()),
-                    right: args[1..].to_vec(),
-                    modifier: None,
-                },
-            };
-            rest = r;
-            continue;
-        }
-        break;
-    }
-    Ok(rest)
-}
-
-/// Build a deferred `Expr::Feed` node, resolving the operator's direction so the
-/// node always means "`source` flows into `sink`". Deferring the fold (rather than
-/// folding into the sink call immediately) lets an assignment/declaration on the
-/// textually-left side split it at Sequencer precedence (looser than `=`); see
-/// `Expr::Feed`, `lower_feed_node`, and `feed_leftmost_operand_mut`.
-fn make_feed_node(op: FeedOp, left: Expr, right: Expr) -> Expr {
-    let (source, sink, append, left_is_source) = match op {
-        FeedOp::ToRight => (left, right, false, true),
-        FeedOp::ToLeft => (right, left, false, false),
-        FeedOp::AppendRight => (left, right, true, true),
-        FeedOp::AppendLeft => (right, left, true, false),
-    };
-    Expr::Feed {
-        source: Box::new(source),
-        sink: Box::new(sink),
-        append,
-        left_is_source,
-    }
-}
-
-/// Lower a (possibly chained) `Expr::Feed` node into its executable form. Nested
-/// feeds on the source/sink sides are lowered first; non-feed operands are left
-/// for the compiler to handle normally. Called by the compiler's `Expr::Feed` arm.
-pub(crate) fn lower_feed_node(source: Expr, sink: Expr, append: bool) -> Expr {
-    let source = match source {
-        Expr::Feed {
-            source: s,
-            sink: k,
-            append: a,
-            ..
-        } => lower_feed_node(*s, *k, a),
-        other => other,
-    };
-    let sink = match sink {
-        Expr::Feed {
-            source: s,
-            sink: k,
-            append: a,
-            ..
-        } => lower_feed_node(*s, *k, a),
-        other => other,
-    };
-    if append {
-        build_append_feed_expr(source, sink)
-    } else {
-        build_pipe_feed_expr(source, sink)
-    }
-}
-
-/// Descend a (possibly chained) `Expr::Feed` to the mutable slot holding its
-/// textually-leftmost operand — the one adjacent to a preceding `=`. For `==>`
-/// chains that is the deepest `source`; for `<==` chains the deepest `sink`.
-/// Used to split `my @a = … ==> …` so the declaration wraps that operand.
-pub(crate) fn feed_leftmost_operand_mut(feed: &mut Expr) -> &mut Expr {
-    match feed {
-        Expr::Feed {
-            source,
-            sink,
-            left_is_source,
-            ..
-        } => {
-            if *left_is_source {
-                feed_leftmost_operand_mut(source)
-            } else {
-                feed_leftmost_operand_mut(sink)
-            }
-        }
-        other => other,
-    }
-}
-
-fn build_pipe_feed_expr(source: Expr, sink: Expr) -> Expr {
-    match sink {
-        // Feeding into a scalar collects the feed into an Array, like an array
-        // sink: `42 ==> $x` and `(1,2,3) ==> $x` both leave `$x` as `[…]`
-        // (an Array), matching Raku (the scalar holds the materialized feed).
-        Expr::Var(name) => Expr::AssignExpr {
-            name,
-            expr: Box::new(Expr::Call {
-                name: Symbol::intern("__mutsu_feed_array_assign"),
-                args: vec![source],
-            }),
-            is_bind: false,
-        },
-        Expr::ArrayVar(name) => Expr::AssignExpr {
-            name: format!("@{}", name),
-            expr: Box::new(Expr::Call {
-                name: Symbol::intern("__mutsu_feed_array_assign"),
-                args: vec![source],
-            }),
-            is_bind: false,
-        },
-        Expr::HashVar(name) => Expr::AssignExpr {
-            name: format!("%{}", name),
-            expr: Box::new(source),
-            is_bind: false,
-        },
-        Expr::CodeVar(name) => Expr::AssignExpr {
-            name: format!("&{}", name),
-            expr: Box::new(source),
-            is_bind: false,
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-        } => Expr::IndexAssign {
-            target,
-            index,
-            value: Box::new(source),
-            is_positional,
-        },
-        Expr::MultiDimIndex { target, dimensions } => Expr::MultiDimIndexAssign {
-            target,
-            dimensions,
-            value: Box::new(source),
-        },
-        Expr::Call { name, mut args } => {
-            args.push(source);
-            Expr::Call { name, args }
-        }
-        Expr::CallOn { target, mut args } => {
-            args.push(source);
-            Expr::CallOn { target, args }
-        }
-        Expr::BareWord(name) => Expr::Call {
-            name: Symbol::intern(&name),
-            args: vec![source],
-        },
-        Expr::Whatever => Expr::Call {
-            name: Symbol::intern("__mutsu_feed_whatever"),
-            args: vec![source],
-        },
-        // Feed into an inline declaration: `... ==> my @o`. The sink parses as a
-        // `DoStmt` wrapping a `VarDecl` with a default (empty) initializer;
-        // rebuild it with the fed source as the initializer so the new variable
-        // (declared in the enclosing scope by `compile_expr_do_stmt`) receives
-        // the feed. Mirrors the bare-variable arms above (an array sink flattens
-        // the source via `__mutsu_feed_array_assign`).
-        Expr::DoStmt(stmt) if matches!(stmt.as_ref(), Stmt::VarDecl { .. }) => {
-            let Stmt::VarDecl {
-                name,
-                type_constraint,
-                is_state,
-                is_our,
-                is_dynamic,
-                is_export,
-                export_tags,
-                custom_traits,
-                where_constraint,
-                ..
-            } = *stmt
-            else {
-                unreachable!("guarded by matches! above")
-            };
-            // Array and scalar sinks materialize the feed into an Array (see the
-            // bare-variable arms above); a hash (`%`) or code (`&`) sink takes
-            // the source directly. Scalar `VarDecl` names are stored without a
-            // sigil (`my $x` -> "x"), arrays/hashes with one ("@o"/"%h").
-            let is_hash_or_code = name.starts_with('%') || name.starts_with('&');
-            let init = if is_hash_or_code {
-                source
-            } else {
-                Expr::Call {
-                    name: Symbol::intern("__mutsu_feed_array_assign"),
-                    args: vec![source],
-                }
-            };
-            Expr::DoStmt(Box::new(Stmt::VarDecl {
-                name,
-                expr: init,
-                type_constraint,
-                is_state,
-                is_our,
-                is_dynamic,
-                is_export,
-                export_tags,
-                custom_traits,
-                where_constraint,
-            }))
-        }
-        other => Expr::CallOn {
-            target: Box::new(other),
-            args: vec![source],
-        },
-    }
-}
-
-fn build_append_feed_expr(source: Expr, sink: Expr) -> Expr {
-    match sink {
-        Expr::Var(name) => Expr::AssignExpr {
-            name: name.clone(),
-            expr: Box::new(Expr::Call {
-                name: Symbol::intern("__mutsu_feed_append"),
-                args: vec![Expr::Var(name), source],
-            }),
-            is_bind: false,
-        },
-        Expr::ArrayVar(name) => Expr::AssignExpr {
-            name: format!("@{}", name.clone()),
-            expr: Box::new(Expr::Call {
-                name: Symbol::intern("__mutsu_feed_append"),
-                args: vec![Expr::ArrayVar(name), source],
-            }),
-            is_bind: false,
-        },
-        Expr::HashVar(name) => Expr::AssignExpr {
-            name: format!("%{}", name.clone()),
-            expr: Box::new(Expr::Call {
-                name: Symbol::intern("__mutsu_feed_append"),
-                args: vec![Expr::HashVar(name), source],
-            }),
-            is_bind: false,
-        },
-        Expr::Index {
-            target,
-            index,
-            is_positional,
-            ..
-        } => {
-            let current = Expr::Index {
-                target: target.clone(),
-                index: index.clone(),
-                is_positional,
-            };
-            Expr::IndexAssign {
-                target,
-                index,
-                value: Box::new(Expr::Call {
-                    name: Symbol::intern("__mutsu_feed_append"),
-                    args: vec![current, source],
-                }),
-                is_positional: true,
-            }
-        }
-        Expr::Whatever => Expr::Call {
-            name: Symbol::intern("__mutsu_feed_append_whatever"),
-            args: vec![source],
-        },
-        other => Expr::Call {
-            name: Symbol::intern("__mutsu_feed_append"),
-            args: vec![other, source],
-        },
-    }
-}
-
-pub(super) fn parse_custom_infix_word(input: &str) -> Option<(String, usize)> {
-    let mut word_match: Option<(String, usize)> = None;
-
-    // Try word-like operators (alphabetic/underscore start)
-    let first = input.chars().next()?;
-    if first.is_alphabetic() || first == '_' {
-        let mut end = first.len_utf8();
-        for ch in input[end..].chars() {
-            if ch.is_alphanumeric() || ch == '_' || ch == '-' {
-                end += ch.len_utf8();
-            } else {
-                break;
-            }
-        }
-        let name = &input[..end];
-        // A declared sigilless term symbol (e.g. `my \term = ...`) is a term, not
-        // an infix operator. Without this guard, `uc term` would be misparsed as
-        // `uc INFIX:<term> ...` instead of the listop call `uc(term)`.
-        let is_declared_term = super::super::stmt::simple::match_user_declared_term_symbol(input)
-            .is_some_and(|(_, consumed, _)| consumed == end);
-        if !is_reserved_infix_word(name) && !is_declared_term {
-            word_match = Some((name.to_string(), end));
-        }
-    }
-
-    // Try user-declared symbol/mixed operators (e.g. ©, ×, _<_)
-    let symbol_match = super::super::stmt::simple::match_user_declared_infix_symbol_op(input);
-
-    // Prefer the longer match
-    match (word_match, symbol_match) {
-        (Some(w), Some(s)) => {
-            if s.1 >= w.1 {
-                Some(s)
-            } else {
-                Some(w)
-            }
-        }
-        (Some(w), None) => Some(w),
-        (None, Some(s)) => Some(s),
-        (None, None) => None,
-    }
-}
-
-fn parse_flipflop_infix(input: &str) -> Option<(String, usize)> {
-    const OPS: &[&str] = &["^fff^", "^fff", "fff^", "fff", "^ff^", "^ff", "ff^", "ff"];
-    for op in OPS {
-        if let Some(rest) = input.strip_prefix(op)
-            && !is_ident_char(rest.as_bytes().first().copied())
-        {
-            return Some(((*op).to_string(), op.len()));
-        }
-    }
-    None
-}
-
-fn is_reserved_infix_word(name: &str) -> bool {
-    if name
-        .chars()
-        .next()
-        .map(|c| c.is_ascii_uppercase())
-        .unwrap_or(false)
-    {
-        return true;
-    }
-    if name.chars().all(|c| c.is_ascii_uppercase()) {
-        return true;
-    }
-    matches!(
-        name,
-        "if" | "unless"
-            | "for"
-            | "while"
-            | "until"
-            | "given"
-            | "when"
-            | "with"
-            | "without"
-            | "my"
-            | "our"
-            | "state"
-            | "has"
-            | "sub"
-            | "method"
-            | "class"
-            | "role"
-            | "grammar"
-            | "module"
-            | "package"
-            | "token"
-            | "rule"
-            | "regex"
-            | "multi"
-            | "proto"
-            | "constant"
-            | "enum"
-            | "subset"
-            | "unit"
-            | "use"
-            | "need"
-            | "return"
-            | "last"
-            | "next"
-            | "redo"
-            | "die"
-            | "fail"
-            | "take"
-            | "do"
-            | "try"
-            | "quietly"
-            | "react"
-            | "whenever"
-            | "loop"
-            | "repeat"
-            | "let"
-            | "temp"
-            | "where"
-            | "is"
-            | "does"
-            | "as"
-            | "of"
-            | "and"
-            | "or"
-            | "not"
-            | "xor"
-            | "andthen"
-            | "orelse"
-            | "notandthen"
-            | "min"
-            | "max"
-            | "cmp"
-            | "coll"
-            | "unicmp"
-            | "leg"
-            | "eq"
-            | "ne"
-            | "lt"
-            | "gt"
-            | "le"
-            | "ge"
-            | "eqv"
-            | "ff"
-            | "fff"
-            | "after"
-            | "before"
-            | "gcd"
-            | "lcm"
-            | "x"
-            | "xx"
-            | "o"
-    )
-}
-
-/// Parse a comma-separated list of range_expr, returning (rest, items).
-fn parse_comma_list_of_range_raw<'a>(input: &'a str) -> PResult<'a, Vec<Expr>> {
-    let (r, first) = range_expr(input)
-        .map_err(|err| enrich_expected_error(err, "expected expression", input.len()))?;
-    let mut items = vec![first];
-    let mut r = r;
-    loop {
-        let (r2, _) = ws(r)?;
-        if !r2.starts_with(',') || r2.starts_with(",,") {
-            break;
-        }
-        let r2 = &r2[1..];
-        let (r2, _) = ws(r2)?;
-        if r2.starts_with(';') || r2.is_empty() || r2.starts_with('}') || r2.starts_with(')') {
-            break;
-        }
-        if let Ok((r3, next)) = range_expr(r2) {
-            items.push(next);
-            r = r3;
-        } else {
-            break;
-        }
-    }
-    Ok((r, items))
-}
-
-/// Parse a comma-separated list of range_expr, returning (rest, Vec<Expr>).
-fn parse_comma_list_of_range<'a>(input: &'a str) -> PResult<'a, Vec<Expr>> {
-    parse_comma_list_of_range_raw(input)
-}
-
-/// Comparison: ==, !=, <, >, <=, >=, eq, ne, lt, gt, le, ge, ~~, !~~, ===, <=>
-fn comparison_expr_mode(input: &str, mode: ExprMode) -> PResult<'_, Expr> {
-    fn regex_rhs_needs_more_parsing(rest: &str) -> bool {
-        let Ok((rest, _)) = ws(rest) else {
-            return false;
-        };
-        parse_junctive_op(rest).is_some() || parse_junction_infix_op(rest).is_some()
-    }
-
-    let (rest, mut left) = structural_comparison_expr_mode(input, mode)?;
-    let (r, _) = ws(rest)?;
-    // Detect Perl 5 =~ and !~ brainos (only when followed by space or m/)
-    if r.starts_with("=~") && !r.starts_with("=~=") && !r.starts_with("=:=") {
-        let after = &r[2..];
-        if after.starts_with(' ')
-            || after.starts_with('\t')
-            || after.starts_with("m/")
-            || after.starts_with("m ")
-        {
-            return Err(PError::fatal(
-                "X::Obsolete: Unsupported use of =~ to do pattern matching; in Raku please use ~~"
-                    .to_string(),
-            ));
-        }
-    }
-    if r.starts_with("!~") && !r.starts_with("!~~") {
-        let after = &r[2..];
-        if after.starts_with(' ')
-            || after.starts_with('\t')
-            || after.starts_with("m/")
-            || after.starts_with("m ")
-        {
-            return Err(PError::fatal(
-                "X::Obsolete: Unsupported use of !~ to do pattern matching; in Raku please use !~~"
-                    .to_string(),
-            ));
-        }
-    }
-    // Detect !% which is an attempt to negate the non-boolean infix:<%;>
-    if r.starts_with("!%") && !r.starts_with("!%%") && !is_ident_char(r.as_bytes().get(2).copied())
-    {
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot negate % because it is not iffy enough",
-            r,
-        ));
-    }
-    // Detect !+ !- !* !/ !~ (non-comparison) which cannot be negated
-    if r.starts_with("!+") && !r.starts_with("!+&") {
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot negate + because it is not iffy enough",
-            r,
-        ));
-    }
-    if r.starts_with("!-") && !r.starts_with("!--") {
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot negate - because it is not iffy enough",
-            r,
-        ));
-    }
-    if r.starts_with("!*") && !r.starts_with("!**") {
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot negate * because it is not iffy enough",
-            r,
-        ));
-    }
-    if r.starts_with("!/") && !r.starts_with("!//") {
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot negate / because it is not iffy enough",
-            r,
-        ));
-    }
-    // Detect !. which is not a valid metaop
-    if r.starts_with("!.") && !r.starts_with("!..") {
-        let after = &r[2..];
-        let after_trimmed = after.trim_start();
-        if after_trimmed.starts_with('"') || after_trimmed.starts_with('\'') {
-            return Err(PError::expected_at(
-                "X::Obsolete: Unsupported use of !. to concatenate strings; in Raku please use ~",
-                r,
-            ));
-        }
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot negate . because it is too fiddly",
-            r,
-        ));
-    }
-    // Detect R. metaop misuse.
-    if r.starts_with("R.") && !r.starts_with("R..") {
-        let after = &r[2..];
-        let after_trimmed = after.trim_start();
-        if after_trimmed.starts_with('"') || after_trimmed.starts_with('\'') {
-            return Err(PError::expected_at(
-                "X::Obsolete: Perl . is dead. Please use ~ to concatenate strings.",
-                r,
-            ));
-        }
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot reverse the args of . because it is too fiddly",
-            r,
-        ));
-    }
-    // Detect X. metaop misuse.
-    if r.starts_with("X.") && !r.starts_with("X..") {
-        let after = &r[2..];
-        let after_trimmed = after.trim_start();
-        if after_trimmed.starts_with('"') || after_trimmed.starts_with('\'') {
-            return Err(PError::expected_at(
-                "X::Obsolete: Perl . is dead. Please use ~ to concatenate strings.",
-                r,
-            ));
-        }
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot do . because it is too fiddly",
-            r,
-        ));
-    }
-    // Detect Z. metaop misuse.
-    if r.starts_with("Z.") && !r.starts_with("Z..") {
-        let after = &r[2..];
-        let after_trimmed = after.trim_start();
-        if after_trimmed.starts_with('"') || after_trimmed.starts_with('\'') {
-            return Err(PError::expected_at(
-                "X::Obsolete: Perl . is dead. Please use ~ to concatenate strings.",
-                r,
-            ));
-        }
-        return Err(PError::expected_at(
-            "X::Syntax::CannotMeta: Cannot do . because it is too fiddly",
-            r,
-        ));
-    }
-    if let Some((op, len)) = parse_negated_meta_comparison_op(r) {
-        let r = &r[len..];
-        let (r, _) = ws(r)?;
-        let (r, right) = if mode == ExprMode::Full {
-            structural_comparison_expr_mode(r, mode).map_err(|err| PError {
-                messages: merge_expected_messages(
-                    "expected expression after comparison operator",
-                    &err.messages,
-                ),
-                remaining_len: err.remaining_len.or(Some(r.len())),
-                exception: None,
-            })?
-        } else {
-            structural_comparison_expr_mode(r, mode)?
-        };
-        let mut result = Expr::Unary {
-            op: TokenKind::Bang,
-            expr: Box::new(Expr::Binary {
-                left: Box::new(left),
-                op: op.token_kind(),
-                right: Box::new(right.clone()),
-            }),
-        };
-        // Support chaining: "a" !after "b" !after "c" → (!after) && (!after)
-        let mut prev_right = right;
-        let mut r = r;
-        loop {
-            let (r2, _) = ws(r)?;
-            // Try negated meta comparison for chaining
-            if let Some((cop, chain_len)) = parse_negated_meta_comparison_op(r2) {
-                let r2 = &r2[chain_len..];
-                let (r2, _) = ws(r2)?;
-                let (r2, next_right) = junctive_expr_mode(r2, mode).map_err(|err| PError {
-                    messages: merge_expected_messages(
-                        "expected expression after chained negated comparison operator",
-                        &err.messages,
-                    ),
-                    remaining_len: err.remaining_len.or(Some(r2.len())),
-                    exception: None,
-                })?;
-                let next_cmp = Expr::Unary {
-                    op: TokenKind::Bang,
-                    expr: Box::new(Expr::Binary {
-                        left: Box::new(prev_right),
-                        op: cop.token_kind(),
-                        right: Box::new(next_right.clone()),
-                    }),
-                };
-                result = Expr::Binary {
-                    left: Box::new(result),
-                    op: TokenKind::AndAnd,
-                    right: Box::new(next_cmp),
-                };
-                prev_right = next_right;
-                r = r2;
-                continue;
-            }
-            // Also try regular comparison for mixed chaining
-            if let Some((cop, chain_len)) = parse_comparison_op(r2) {
-                let r2 = &r2[chain_len..];
-                let (r2, _) = ws(r2)?;
-                let (r2, next_right) = junctive_expr_mode(r2, mode).map_err(|err| PError {
-                    messages: merge_expected_messages(
-                        "expected expression after chained comparison operator",
-                        &err.messages,
-                    ),
-                    remaining_len: err.remaining_len.or(Some(r2.len())),
-                    exception: None,
-                })?;
-                let next_cmp = Expr::Binary {
-                    left: Box::new(prev_right),
-                    op: cop.token_kind(),
-                    right: Box::new(next_right.clone()),
-                };
-                result = Expr::Binary {
-                    left: Box::new(result),
-                    op: TokenKind::AndAnd,
-                    right: Box::new(next_cmp),
-                };
-                prev_right = next_right;
-                r = r2;
-                continue;
-            }
-            break;
-        }
-        return Ok((r, result));
-    }
-    if let Some((op, len)) = parse_comparison_op(r) {
-        let r = &r[len..];
-        let (r, _) = ws(r)?;
-        // Track whether the smartmatch RHS is a regex literal.  When it is,
-        // chaining with subsequent comparison operators (eq, ==, …) must be
-        // suppressed because the regex literal terminates the smartmatch RHS
-        // at its closing delimiter, so `X ~~ /pat/ eq Z` should parse as
-        // `(X ~~ /pat/) eq Z` (left-associative), not as a chained comparison.
-        let mut rhs_is_regex_lit = false;
-        let (r, mut right) = if matches!(op, ComparisonOp::SmartMatch | ComparisonOp::SmartNotMatch)
-        {
-            if let Ok((rest, expr)) = crate::parser::primary::regex::regex_lit(r) {
-                if regex_rhs_needs_more_parsing(rest) {
-                    if mode == ExprMode::Full {
-                        structural_comparison_expr_mode(r, mode).map_err(|err| {
-                            // Preserve fatal errors with structured exceptions (e.g., X::Obsolete)
-                            if err.is_fatal() && err.exception.is_some() {
-                                return err;
-                            }
-                            PError {
-                                messages: merge_expected_messages(
-                                    "expected expression after comparison operator",
-                                    &err.messages,
-                                ),
-                                remaining_len: err.remaining_len.or(Some(r.len())),
-                                exception: None,
-                            }
-                        })?
-                    } else {
-                        structural_comparison_expr_mode(r, mode)?
-                    }
-                } else {
-                    rhs_is_regex_lit = true;
-                    (rest, expr)
-                }
-            } else if mode == ExprMode::Full {
-                structural_comparison_expr_mode(r, mode).map_err(|err| {
-                    // Preserve fatal errors with structured exceptions (e.g., X::Obsolete)
-                    if err.is_fatal() && err.exception.is_some() {
-                        return err;
-                    }
-                    PError {
-                        messages: merge_expected_messages(
-                            "expected expression after comparison operator",
-                            &err.messages,
-                        ),
-                        remaining_len: err.remaining_len.or(Some(r.len())),
-                        exception: None,
-                    }
-                })?
-            } else {
-                structural_comparison_expr_mode(r, mode)?
-            }
-        } else if mode == ExprMode::Full {
-            structural_comparison_expr_mode(r, mode).map_err(|err| {
-                // Preserve fatal errors with structured exceptions (e.g., X::Obsolete)
-                if err.is_fatal() && err.exception.is_some() {
-                    return err;
-                }
-                PError {
-                    messages: merge_expected_messages(
-                        "expected expression after comparison operator",
-                        &err.messages,
-                    ),
-                    remaining_len: err.remaining_len.or(Some(r.len())),
-                    exception: None,
-                }
-            })?
-        } else {
-            structural_comparison_expr_mode(r, mode)?
-        };
-        if matches!(op, ComparisonOp::SmartMatch | ComparisonOp::SmartNotMatch) {
-            right = wrap_smartmatch_rhs(right);
-            // Wrap LHS WhateverCode subexpressions (e.g. `*.abs ~~ Code`).
-            // SmartMatch suppresses top-level WhateverCode wrapping, but the LHS
-            // should still be curried when it contains Whatever.
-            if super::should_wrap_whatevercode(&left) {
-                left = super::wrap_whatevercode(&left);
-            }
-            // Bare `* ~~ Type` curries to WhateverCode `{ $_ ~~ Type }`.
-            // This is only done here (not in should_curry_whatever) because at the
-            // top-level wrapping stage, parenthesized `((*)) ~~ Type` also has
-            // Expr::Whatever as LHS and should NOT curry — only a true bare `*`
-            // operand in the precedence parser should trigger this.
-            // We distinguish bare `*` from `((*))` by checking whether the consumed
-            // input for the LHS started with `(`.
-            if matches!(&left, Expr::Whatever) {
-                let lhs_text = input[..input.len() - rest.len()].trim_start();
-                if !lhs_text.starts_with('(') {
-                    let sm_expr = Expr::Binary {
-                        left: Box::new(Expr::Var("_".to_string())),
-                        op: op.token_kind(),
-                        right: Box::new(right),
-                    };
-                    return Ok((
-                        r,
-                        Expr::Lambda {
-                            param: "_".to_string(),
-                            body: vec![crate::ast::Stmt::Expr(sm_expr)],
-                            is_whatever_code: true,
-                        },
-                    ));
-                }
-            }
-        }
-        // When the smartmatch RHS is a regex literal, do not chain with
-        // subsequent comparison operators.  The regex literal terminates the
-        // smartmatch and the following operator applies to the match result.
-        if rhs_is_regex_lit {
-            return Ok((r, make_chain_cmp(left, op.token_kind(), right, false)));
-        }
-        let mut operands = vec![left, right];
-        let mut chain_ops: Vec<(TokenKind, bool)> = vec![(op.token_kind(), false)];
-        let mut r = r;
-        loop {
-            let (r2, _) = ws(r)?;
-            if let Some((cop, chain_len)) = parse_negated_meta_comparison_op(r2) {
-                if is_structural_comparison_op(cop) {
-                    break;
-                }
-                let r2 = &r2[chain_len..];
-                let (r2, _) = ws(r2)?;
-                let (r2, next_right) =
-                    structural_comparison_expr_mode(r2, mode).map_err(|err| PError {
-                        messages: merge_expected_messages(
-                            "expected expression after chained negated comparison operator",
-                            &err.messages,
-                        ),
-                        remaining_len: err.remaining_len.or(Some(r2.len())),
-                        exception: None,
-                    })?;
-                if let (Some(prev), Some(next)) = (
-                    comparison_nonassoc_key(&chain_ops.last().expect("chain op").0),
-                    comparison_nonassoc_key(&cop.token_kind()),
-                ) {
-                    return Err(non_associative_error(&format!("{prev} and {next}")));
-                }
-                chain_ops.push((cop.token_kind(), true));
-                operands.push(next_right);
-                r = r2;
-                continue;
-            }
-            if let Some((cop, chain_len)) = parse_comparison_op(r2) {
-                if is_structural_comparison_op(cop) {
-                    break;
-                }
-                let r2 = &r2[chain_len..];
-                let (r2, _) = ws(r2)?;
-                let (r2, next_right) =
-                    structural_comparison_expr_mode(r2, mode).map_err(|err| PError {
-                        messages: merge_expected_messages(
-                            "expected expression after chained comparison operator",
-                            &err.messages,
-                        ),
-                        remaining_len: err.remaining_len.or(Some(r2.len())),
-                        exception: None,
-                    })?;
-                if let (Some(prev), Some(next)) = (
-                    comparison_nonassoc_key(&chain_ops.last().expect("chain op").0),
-                    comparison_nonassoc_key(&cop.token_kind()),
-                ) {
-                    return Err(non_associative_error(&format!("{prev} and {next}")));
-                }
-                let next_right =
-                    if matches!(cop, ComparisonOp::SmartMatch | ComparisonOp::SmartNotMatch) {
-                        wrap_smartmatch_rhs(next_right)
-                    } else {
-                        next_right
-                    };
-                chain_ops.push((cop.token_kind(), false));
-                operands.push(next_right);
-                r = r2;
-                continue;
-            }
-            break;
-        }
-        if chain_ops.len() == 1 {
-            return Ok((
-                r,
-                make_chain_cmp(
-                    operands[0].clone(),
-                    chain_ops[0].0.clone(),
-                    operands[1].clone(),
-                    false,
-                ),
-            ));
-        }
-        let result = if operands.iter().any(contains_whatever) {
-            // Keep the historical `lhs < * && * < rhs` shape so WhateverCode wrapping
-            // can preserve placeholder semantics in expression arguments.
-            build_chain_cmp_expr_with_repeated_middle(&operands, &chain_ops)
-        } else {
-            build_chain_cmp_expr(&operands, &chain_ops, 0, operands[0].clone())
-        };
-        return Ok((r, result));
-    }
-
-    Ok((rest, left))
-}
-
-fn make_chain_cmp(left: Expr, op: TokenKind, right: Expr, negated: bool) -> Expr {
-    let cmp = Expr::Binary {
-        left: Box::new(left),
-        op,
-        right: Box::new(right),
-    };
-    if negated {
-        Expr::Unary {
-            op: TokenKind::Bang,
-            expr: Box::new(cmp),
-        }
-    } else {
-        cmp
-    }
-}
-
-fn build_chain_cmp_expr(
-    operands: &[Expr],
-    ops: &[(TokenKind, bool)],
-    index: usize,
-    left: Expr,
-) -> Expr {
-    let (op, negated) = ops[index].clone();
-    if index == ops.len() - 1 {
-        return make_chain_cmp(left, op, operands[index + 1].clone(), negated);
-    }
-
-    let tmp_idx = CHAIN_CMP_TMP_COUNTER.fetch_add(1, Ordering::Relaxed);
-    let tmp_name = format!("__mutsu_chain_cmp_{tmp_idx}");
-    let tmp_var = Expr::Var(tmp_name.clone());
-    let cmp = make_chain_cmp(left, op, tmp_var.clone(), negated);
-    let rest = build_chain_cmp_expr(operands, ops, index + 1, tmp_var.clone());
-    Expr::DoBlock {
-        body: vec![
-            Stmt::VarDecl {
-                name: tmp_name,
-                expr: operands[index + 1].clone(),
-                type_constraint: None,
-                is_state: false,
-                is_our: false,
-                is_dynamic: false,
-                is_export: false,
-                export_tags: Vec::new(),
-                custom_traits: Vec::new(),
-                where_constraint: None,
-            },
-            Stmt::Expr(Expr::Binary {
-                left: Box::new(cmp),
-                op: TokenKind::AndAnd,
-                right: Box::new(rest),
-            }),
-        ],
-        label: None,
-    }
-}
-
-fn build_chain_cmp_expr_with_repeated_middle(operands: &[Expr], ops: &[(TokenKind, bool)]) -> Expr {
-    let mut result = make_chain_cmp(
-        operands[0].clone(),
-        ops[0].0.clone(),
-        operands[1].clone(),
-        ops[0].1,
-    );
-    let mut prev_right = operands[1].clone();
-    for (idx, (op, negated)) in ops.iter().enumerate().skip(1) {
-        let next_right = operands[idx + 1].clone();
-        let next_cmp = make_chain_cmp(prev_right, op.clone(), next_right.clone(), *negated);
-        result = Expr::Binary {
-            left: Box::new(result),
-            op: TokenKind::AndAnd,
-            right: Box::new(next_cmp),
-        };
-        prev_right = next_right;
-    }
-    result
-}
-
-fn wrap_smartmatch_rhs(right: Expr) -> Expr {
-    match right {
-        // Keep Pair shape for `%hash ~~ key => !*.foo` by wrapping only value side.
-        Expr::Binary {
-            left,
-            op: TokenKind::FatArrow,
-            right,
-        } => {
-            let value = if contains_whatever(&right) && !matches!(&*right, Expr::Whatever) {
-                wrap_whatevercode(&right)
-            } else {
-                *right
-            };
-            Expr::Binary {
-                left,
-                op: TokenKind::FatArrow,
-                right: Box::new(value),
-            }
-        }
-        other => {
-            if contains_whatever(&other) && !matches!(&other, Expr::Whatever) {
-                wrap_whatevercode(&other)
-            } else {
-                other
-            }
-        }
-    }
-}
-
-/// Extract a comparison operator from the start of the input, returning the op and its length.
-fn parse_comparison_op(r: &str) -> Option<(ComparisonOp, usize)> {
-    // Unicode comparison operators
-    if r.starts_with('\u{2A75}') {
-        // ⩵ (U+2A75) — numeric equality (alias for ==)
-        return Some((ComparisonOp::NumEq, '\u{2A75}'.len_utf8()));
-    } else if r.starts_with('\u{2A76}') {
-        // ⩶ (U+2A76) — value identity (alias for ===)
-        return Some((ComparisonOp::StrictEq, '\u{2A76}'.len_utf8()));
-    } else if r.starts_with('\u{2260}') {
-        // ≠ (U+2260) — numeric inequality (alias for !=)
-        return Some((ComparisonOp::NumNe, '\u{2260}'.len_utf8()));
-    } else if r.starts_with('\u{2264}') {
-        // ≤ (U+2264) — numeric less-than-or-equal (alias for <=)
-        return Some((ComparisonOp::NumLe, '\u{2264}'.len_utf8()));
-    } else if r.starts_with('\u{2265}') {
-        // ≥ (U+2265) — numeric greater-than-or-equal (alias for >=)
-        return Some((ComparisonOp::NumGe, '\u{2265}'.len_utf8()));
-    }
-    // ≅ (U+2245) — approximately equal
-    if r.starts_with('\u{2245}') {
-        return Some((ComparisonOp::ApproxEq, '\u{2245}'.len_utf8()));
-    }
-    if r.starts_with("=~=") {
-        return Some((ComparisonOp::ApproxEq, 3));
-    }
-    if r.starts_with("!=:=") {
-        return Some((ComparisonOp::ContainerNe, 4));
-    }
-    if r.starts_with("=:=") {
-        return Some((ComparisonOp::ContainerEq, 3));
-    }
-    if r.starts_with("===") {
-        Some((ComparisonOp::StrictEq, 3))
-    } else if r.starts_with("==") && !r.starts_with("===") && !r.starts_with("==>") {
-        Some((ComparisonOp::NumEq, 2))
-    } else if r.starts_with("!%%") {
-        Some((ComparisonOp::NotDivisibleBy, 3))
-    } else if r.starts_with("!===") {
-        Some((ComparisonOp::StrictNe, 4))
-    } else if r.starts_with("!=") {
-        Some((ComparisonOp::NumNe, 2))
-    } else if r.starts_with("!~~") {
-        Some((ComparisonOp::SmartNotMatch, 3))
-    } else if r.starts_with("~~") {
-        Some((ComparisonOp::SmartMatch, 2))
-    } else if r.starts_with("<=>") {
-        Some((ComparisonOp::Spaceship, 3))
-    } else if r.starts_with("<=") && !r.starts_with("<=>") && !r.starts_with("<==") {
-        Some((ComparisonOp::NumLe, 2))
-    } else if r.starts_with(">=") {
-        Some((ComparisonOp::NumGe, 2))
-    } else if r.starts_with('<') && !r.starts_with("<<") && !r.starts_with("<=") {
-        Some((ComparisonOp::NumLt, 1))
-    } else if r.starts_with('>') && !r.starts_with(">>") && !r.starts_with(">=") {
-        Some((ComparisonOp::NumGt, 1))
-    } else if r.starts_with("eq") && !is_ident_char(r.as_bytes().get(2).copied()) {
-        Some((ComparisonOp::StrEq, 2))
-    } else if r.starts_with("ne") && !is_ident_char(r.as_bytes().get(2).copied()) {
-        Some((ComparisonOp::StrNe, 2))
-    } else if r.starts_with("lt") && !is_ident_char(r.as_bytes().get(2).copied()) {
-        Some((ComparisonOp::StrLt, 2))
-    } else if r.starts_with("gt") && !is_ident_char(r.as_bytes().get(2).copied()) {
-        Some((ComparisonOp::StrGt, 2))
-    } else if r.starts_with("le") && !is_ident_char(r.as_bytes().get(2).copied()) {
-        Some((ComparisonOp::StrLe, 2))
-    } else if r.starts_with("ge") && !is_ident_char(r.as_bytes().get(2).copied()) {
-        Some((ComparisonOp::StrGe, 2))
-    } else if r.starts_with("leg") && !is_ident_char(r.as_bytes().get(3).copied()) {
-        Some((ComparisonOp::Leg, 3))
-    } else if r.starts_with("cmp") && !is_ident_char(r.as_bytes().get(3).copied()) {
-        Some((ComparisonOp::Cmp, 3))
-    } else if r.starts_with("coll") && !is_ident_char(r.as_bytes().get(4).copied()) {
-        Some((ComparisonOp::Coll, 4))
-    } else if r.starts_with("unicmp") && !is_ident_char(r.as_bytes().get(6).copied()) {
-        Some((ComparisonOp::Unicmp, 6))
-    } else if r.starts_with("eqv") && !is_ident_char(r.as_bytes().get(3).copied()) {
-        Some((ComparisonOp::Eqv, 3))
-    } else if r.starts_with("before") && !is_ident_char(r.as_bytes().get(6).copied()) {
-        Some((ComparisonOp::Before, 6))
-    } else if r.starts_with("after") && !is_ident_char(r.as_bytes().get(5).copied()) {
-        Some((ComparisonOp::After, 5))
-    } else {
-        None
-    }
-}
-
-fn parse_negated_meta_comparison_op(r: &str) -> Option<(ComparisonOp, usize)> {
-    let inner = r.strip_prefix('!')?;
-    let (op, len) = parse_comparison_op(inner)?;
-    // Operators that already have their own !-prefixed spelling are not meta-negated forms.
-    if matches!(
-        op,
-        ComparisonOp::NumNe
-            | ComparisonOp::NotDivisibleBy
-            | ComparisonOp::SmartMatch
-            | ComparisonOp::SmartNotMatch
-    ) {
-        return None;
-    }
-    Some((op, len + 1))
-}
-
-/// Range: ..  ..^  ^..  ^..^
-pub(super) fn range_expr(input: &str) -> PResult<'_, Expr> {
-    let (rest, left) = structural_expr(input)?;
-    let (r, _) = ws(rest)?;
-
-    if let Some(stripped) = r.strip_prefix("^..^") {
-        check_range_precedence_worry(input)?;
-        let (r, _) = ws(stripped)?;
-        let (r, right) = structural_expr(r).map_err(|err| {
-            enrich_expected_error(err, "expected range RHS after '^..^'", r.len())
-        })?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(left),
-                op: TokenKind::CaretDotDotCaret,
-                right: Box::new(right),
-            },
-        ));
-    }
-    if let Some(stripped) = r.strip_prefix("^..") {
-        check_range_precedence_worry(input)?;
-        let (r, _) = ws(stripped)?;
-        let (r, right) = structural_expr(r)
-            .map_err(|err| enrich_expected_error(err, "expected range RHS after '^..'", r.len()))?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(left),
-                op: TokenKind::CaretDotDot,
-                right: Box::new(right),
-            },
-        ));
-    }
-    if let Some(stripped) = r.strip_prefix("..^") {
-        check_range_precedence_worry(input)?;
-        let (r, _) = ws(stripped)?;
-        let (r, right) = structural_expr(r)
-            .map_err(|err| enrich_expected_error(err, "expected range RHS after '..^'", r.len()))?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(left),
-                op: TokenKind::DotDotCaret,
-                right: Box::new(right),
-            },
-        ));
-    }
-    if r.starts_with("..") && !r.starts_with("...") {
-        check_range_precedence_worry(input)?;
-        let r = &r[2..];
-        let (r, _) = ws(r)?;
-        let (r, right) = structural_expr(r)
-            .map_err(|err| enrich_expected_error(err, "expected range RHS after '..'", r.len()))?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(left),
-                op: TokenKind::DotDot,
-                right: Box::new(right),
-            },
-        ));
-    }
-    Ok((rest, left))
-}
+mod assign;
+mod chain_cmp;
+mod comparison;
+mod custom_infix;
+mod errors;
+mod feed;
+mod list_infix;
+mod list_infix_loop;
+mod logic;
+mod logic2;
+mod operands;
+mod ternary;
+
+// ---- Re-exports preserving each public function's original visibility ----
+pub(super) use chain_cmp::range_expr;
+pub(super) use custom_infix::parse_custom_infix_word;
+pub(super) use operands::{loose_prefix_operand, multiplicative_operand, power_operand};
+pub(super) use ternary::{ternary, ternary_mode, ternary_no_assign};
+// These take `pub(super)` types (ComparisonOp/FeedOp), so re-export at the same
+// visibility to avoid leaking a more-private type (private_interfaces).
+pub(super) use chain_cmp::{parse_comparison_op, parse_negated_meta_comparison_op};
+pub(super) use feed::make_feed_node;
+pub(super) use ternary::is_structural_comparison_op;
+
+pub(in crate::parser) use ternary::call_arg_expr;
+
+pub(crate) use feed::{feed_leftmost_operand_mut, lower_feed_node};
+
+// ---- Internal cross-submodule helper re-exports (so `use super::*` reaches them) ----
+pub(crate) use assign::{
+    assign_to_target_expr, build_compound_assign_target_expr, list_lvalue_assign_expr,
+    parse_assignment_rhs_mode, unwrap_grouped_lvalue,
+};
+pub(crate) use chain_cmp::{
+    build_chain_cmp_expr, build_chain_cmp_expr_with_repeated_middle, make_chain_cmp,
+    wrap_smartmatch_rhs,
+};
+pub(crate) use comparison::comparison_expr_mode;
+pub(crate) use custom_infix::{
+    parse_comma_list_of_range, parse_comma_list_of_range_raw, parse_flipflop_infix,
+};
+pub(crate) use errors::{
+    check_range_precedence_worry, conditional_precedence_too_loose_error, non_associative_error,
+    non_associative_pair_error, non_list_associative_error, syntax_exception,
+};
+pub(crate) use list_infix::{list_infix_expr, sequence_expr};
+pub(crate) use list_infix_loop::parse_list_infix_loop;
+pub(crate) use logic::{assign_not_expr_mode, or_expr_mode, or_expr_no_assign_mode};
+pub(crate) use logic2::{junctive_expr_mode, not_expr_mode, or_or_expr_mode};
+pub(crate) use ternary::{comparison_nonassoc_key, structural_comparison_expr_mode};
 
 #[cfg(test)]
 mod tests {
